Get a tidy dataset of NHL goalies game logs

The function tidy_goalies_gamelogs() is meant to be a user-friendly way of getting the NHL game logs of selected goalies.

tidy_goalies_gamelogs(
  players_id,
  seasons_id = NULL,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)

Arguments

players_id
Integer vector indicating the NHL ID of goalies for whom the game logs will be returned.

seasons_id
(optional) Character vector indicating the seasons for which the game logs will be returned. The required format for each character is 'xxxxyyyy'. Default to NULL which will return the full career game logs for each player.

regular
(optional) Logical indicating if the regular season game logs should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs game logs should be returned. Default to TRUE.

tz
(optional) Character specifying the timezone that should be used for datetime. Default to the user system timezone.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default can be set globally with options("tidynhl.data.table").

Examples

# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get the NHL career game logs of Roberto Luongo
tidy_goalies_gamelogs(8466141L)

#>          player_name season_years season_type       game_datetime                        venue_name team_status team_abbreviation team_score opponent_score opponent_abbreviation game_status game_nbot game_shootout goalie_games goalie_started goalie_wins goalie_losses goalie_ties goalie_ot goalie_shutouts goalie_shotagainst goalie_goalagainst goalie_savepct goalie_gaa goalie_toi goalie_ev_shotagainst goalie_ev_goalagainst goalie_ev_savepct goalie_pp_shotagainst goalie_pp_goalagainst goalie_pp_savepct goalie_pk_shotagainst goalie_pk_goalagainst goalie_pk_savepct
#>    1: Roberto Luongo      1999-00     regular 1999-11-29 00:00:00                       FleetCenter        away               NYI          2              1                   BOS       final         0         FALSE            1              1           1             0           0         0               0                 44                  1      0.9772727  1.0000000   60.00000                    32                     1         0.9687500                    11                     0         1.0000000                     1                     0                 1
#>    2: Roberto Luongo      1999-00     regular 1999-12-01 00:30:00 Nassau Veterans Memorial Coliseum        home               NYI          1              2                   DAL       final         0         FALSE            1              1           0             1           0         0               0                 29                  2      0.9310345  2.0270270   59.20000                    24                     0         1.0000000                     3                     2         0.3333333                     2                     0                 1
#>    3: Roberto Luongo      1999-00     regular 1999-12-08 00:00:00                        MCI Center        away               NYI          2              4                   WSH       final         0         FALSE            1              1           0             1           0         0               0                 26                  3      0.8846154  3.0184461   59.63333                    15                     2         0.8666667                    11                     1         0.9090909                     0                     0               NaN
#>    4: Roberto Luongo      1999-00     regular 1999-12-11 20:00:00                      Corel Centre        away               NYI          1              1                   OTT       final         1         FALSE            1              1           0             0           1         0               0                 37                  1      0.9729730  0.9230769   65.00000                    34                     1         0.9705882                     3                     0         1.0000000                     0                     0               NaN
#>    5: Roberto Luongo      1999-00     regular 1999-12-15 00:30:00 Nassau Veterans Memorial Coliseum        home               NYI          2              4                   EDM       final         0         FALSE            1              1           0             1           0         0               0                 31                  4      0.8709677  4.0000000   60.00000                    28                     3         0.8928571                     2                     1         0.5000000                     1                     0                 1
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> 1110: Roberto Luongo      2018-19     regular 2019-03-25 23:00:00                  Scotiabank Arena        away               FLA          5              7                   TOR       final         0         FALSE            1              0           0             1           0         0               0                 33                  5      0.8484848  5.6074766   53.50000                    27                     4         0.8518519                     5                     1         0.8000000                     1                     0                 1
#> 1111: Roberto Luongo      2018-19     regular 2019-03-28 23:30:00              Canadian Tire Centre        away               FLA          5              2                   OTT       final         0         FALSE            1              1           1             0           0         0               0                 30                  2      0.9333333  2.0000000   60.00000                    27                     2         0.9259259                     2                     0         1.0000000                     1                     0                 1
#> 1112: Roberto Luongo      2018-19     regular 2019-03-30 17:00:00                         TD Garden        away               FLA          4              1                   BOS       final         0         FALSE            1              1           1             0           0         0               0                 31                  1      0.9677419  1.0000000   60.00000                    30                     1         0.9666667                     1                     0         1.0000000                     0                     0               NaN
#> 1113: Roberto Luongo      2018-19     regular 2019-04-01 23:00:00                       BB&T Center        home               FLA          5              3                   WSH       final         0         FALSE            1              1           1             0           0         0               0                 30                  3      0.9000000  3.0000000   60.00000                    30                     3         0.9000000                     0                     0               NaN                     0                     0               NaN
#> 1114: Roberto Luongo      2018-19     regular 2019-04-06 23:00:00                       BB&T Center        home               FLA          3              4                   NJD       final         1         FALSE            1              1           0             0           0         1               0                 27                  4      0.8518519  3.9658496   60.51667                    20                     3         0.8500000                     6                     1         0.8333333                     1                     0                 1

# Get the 2010-11 NHL playoffs game logs of both Roberto Luongo and Corey Crawford, keeping the
# IDs
tidy_goalies_gamelogs(
  players_id = c(8466141L, 8470645L),
  seasons_id = "20102011",
  regular = FALSE,
  keep_id = TRUE
)

#>     player_id    player_name season_id season_years season_type    game_id       game_datetime              venue_name team_status team_id team_abbreviation team_score opponent_score opponent_abbreviation opponent_id game_status game_nbot game_shootout goalie_games goalie_started goalie_wins goalie_losses goalie_ties goalie_ot goalie_shutouts goalie_shotagainst goalie_goalagainst goalie_savepct goalie_gaa goalie_toi goalie_ev_shotagainst goalie_ev_goalagainst goalie_ev_savepct goalie_pp_shotagainst goalie_pp_goalagainst goalie_pp_savepct goalie_pk_shotagainst goalie_pk_goalagainst goalie_pk_savepct
#>  1:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030151 2011-04-14 02:00:00            Rogers Arena        home      23               VAN          2              0                   CHI          16       final         0         FALSE            1              1           1             0           0         0               1                 32                  0      1.0000000   0.000000  60.000000                    28                     0         1.0000000                     4                     0         1.0000000                     0                     0               NaN
#>  2:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030152 2011-04-16 02:00:00            Rogers Arena        home      23               VAN          4              3                   CHI          16       final         0         FALSE            1              1           1             0           0         0               0                 26                  3      0.8846154   3.000000  60.000000                    26                     3         0.8846154                     0                     0               NaN                     0                     0               NaN
#>  3:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030153 2011-04-18 00:00:00           United Center        away      23               VAN          3              2                   CHI          16       final         0         FALSE            1              1           1             0           0         0               0                 32                  2      0.9375000   2.000000  60.000000                    23                     0         1.0000000                     9                     2         0.7777778                     0                     0               NaN
#>  4:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030154 2011-04-20 00:00:00           United Center        away      23               VAN          2              7                   CHI          16       final         0         FALSE            1              1           0             1           0         0               0                 28                  6      0.7857143   8.172531  44.050000                    22                     5         0.7727273                     6                     1         0.8333333                     0                     0               NaN
#>  5:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030155 2011-04-22 02:00:00            Rogers Arena        home      23               VAN          0              5                   CHI          16       final         0         FALSE            1              1           0             1           0         0               0                 12                  4      0.6666667  11.294118  21.250000                     8                     2         0.7500000                     3                     2         0.3333333                     1                     0         1.0000000
#>  6:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030156 2011-04-24 23:30:00           United Center        away      23               VAN          3              4                   CHI          16       final         1         FALSE            1              0           0             1           0         0               0                 13                  1      0.9230769   1.819101  32.983333                    13                     1         0.9230769                     0                     0               NaN                     0                     0               NaN
#>  7:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030157 2011-04-27 02:00:00            Rogers Arena        home      23               VAN          2              1                   CHI          16       final         1         FALSE            1              1           1             0           0         0               0                 32                  1      0.9687500   0.917899  65.366667                    29                     0         1.0000000                     1                     0         1.0000000                     2                     1         0.5000000
#>  8:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030231 2011-04-29 01:00:00            Rogers Arena        home      23               VAN          1              0                   NSH          18       final         0         FALSE            1              1           1             0           0         0               1                 20                  0      1.0000000   0.000000  59.966667                    15                     0         1.0000000                     4                     0         1.0000000                     1                     0         1.0000000
#>  9:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030232 2011-05-01 01:00:00            Rogers Arena        home      23               VAN          1              2                   NSH          18       final         2         FALSE            1              1           0             1           0         0               0                 46                  2      0.9565217   1.265156  94.850000                    39                     2         0.9487179                     7                     0         1.0000000                     0                     0               NaN
#> 10:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030233 2011-05-04 01:00:00       Bridgestone Arena        away      23               VAN          3              2                   NSH          18       final         1         FALSE            1              1           1             0           0         0               0                 30                  2      0.9333333   1.696113  70.750000                    25                     1         0.9600000                     3                     0         1.0000000                     2                     1         0.5000000
#> 11:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030234 2011-05-06 00:30:00       Bridgestone Arena        away      23               VAN          4              2                   NSH          18       final         0         FALSE            1              1           1             0           0         0               0                 21                  2      0.9047619   2.000556  59.983333                    15                     1         0.9333333                     6                     1         0.8333333                     0                     0               NaN
#> 12:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030235 2011-05-08 00:00:00            Rogers Arena        home      23               VAN          3              4                   NSH          18       final         0         FALSE            1              1           0             1           0         0               0                 23                  4      0.8260870   4.066648  59.016667                    19                     3         0.8421053                     2                     0         1.0000000                     2                     1         0.5000000
#> 13:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030236 2011-05-10 00:00:00       Bridgestone Arena        away      23               VAN          2              1                   NSH          18       final         0         FALSE            1              1           1             0           0         0               0                 24                  1      0.9583333   1.003904  59.766667                    15                     1         0.9333333                     7                     0         1.0000000                     2                     0         1.0000000
#> 14:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030321 2011-05-16 00:00:00            Rogers Arena        home      23               VAN          3              2                   SJS          28       final         0         FALSE            1              1           1             0           0         0               0                 29                  2      0.9310345   2.000000  60.000000                    28                     1         0.9642857                     1                     1         0.0000000                     0                     0               NaN
#> 15:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030322 2011-05-19 01:00:00            Rogers Arena        home      23               VAN          7              3                   SJS          28       final         0         FALSE            1              1           1             0           0         0               0                 31                  3      0.9032258   3.011712  59.766667                    26                     1         0.9615385                     3                     2         0.3333333                     2                     0         1.0000000
#> 16:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030323 2011-05-21 01:00:00 HP Pavilion at San Jose        away      23               VAN          3              4                   SJS          28       final         0         FALSE            1              1           0             1           0         0               0                 38                  4      0.8947368   4.087425  58.716667                    18                     1         0.9444444                    19                     3         0.8421053                     1                     0         1.0000000
#> 17:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030324 2011-05-22 19:00:00 HP Pavilion at San Jose        away      23               VAN          4              2                   SJS          28       final         0         FALSE            1              1           1             0           0         0               0                 35                  2      0.9428571   2.000000  60.000000                    26                     2         0.9230769                     8                     0         1.0000000                     1                     0         1.0000000
#> 18:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030325 2011-05-25 01:00:00            Rogers Arena        home      23               VAN          3              2                   SJS          28       final         2         FALSE            1              1           1             0           0         0               0                 56                  2      0.9642857   1.343785  89.300000                    46                     1         0.9782609                     9                     1         0.8888889                     1                     0         1.0000000
#> 19:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030411 2011-06-02 00:00:00            Rogers Arena        home      23               VAN          1              0                   BOS           6       final         0         FALSE            1              1           1             0           0         0               1                 36                  0      1.0000000   0.000000  59.766667                    22                     0         1.0000000                    12                     0         1.0000000                     2                     0         1.0000000
#> 20:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030412 2011-06-05 00:00:00            Rogers Arena        home      23               VAN          3              2                   BOS           6       final         1         FALSE            1              1           1             0           0         0               0                 30                  2      0.9333333   1.993908  60.183333                    23                     1         0.9565217                     7                     1         0.8571429                     0                     0               NaN
#> 21:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030413 2011-06-07 00:00:00               TD Garden        away      23               VAN          1              8                   BOS           6       final         0         FALSE            1              1           0             1           0         0               0                 38                  8      0.7894737   8.000000  60.000000                    18                     4         0.7777778                    13                     2         0.8461538                     7                     2         0.7142857
#> 22:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030414 2011-06-09 00:00:00               TD Garden        away      23               VAN          0              4                   BOS           6       final         0         FALSE            1              1           0             1           0         0               0                 20                  4      0.8000000   5.544859  43.283333                    17                     4         0.7647059                     3                     0         1.0000000                     0                     0               NaN
#> 23:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030415 2011-06-11 00:00:00            Rogers Arena        home      23               VAN          1              0                   BOS           6       final         0         FALSE            1              1           1             0           0         0               1                 31                  0      1.0000000   0.000000  60.000000                    24                     0         1.0000000                     7                     0         1.0000000                     0                     0               NaN
#> 24:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030416 2011-06-14 00:00:00               TD Garden        away      23               VAN          2              5                   BOS           6       final         0         FALSE            1              1           0             1           0         0               0                  8                  3      0.6250000  20.970874   8.583333                     6                     2         0.6666667                     2                     1         0.5000000                     0                     0               NaN
#> 25:   8466141 Roberto Luongo  20102011      2010-11    playoffs 2010030417 2011-06-16 00:00:00            Rogers Arena        home      23               VAN          0              4                   BOS           6       final         0         FALSE            1              1           0             1           0         0               0                 20                  3      0.8500000   3.020979  59.583333                    17                     2         0.8823529                     1                     0         1.0000000                     2                     1         0.5000000
#> 26:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030151 2011-04-14 02:00:00            Rogers Arena        away      16               CHI          0              2                   VAN          23       final         0         FALSE            1              1           0             1           0         0               0                 33                  2      0.9393939   2.068966  58.000000                    29                     2         0.9310345                     2                     0         1.0000000                     2                     0         1.0000000
#> 27:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030152 2011-04-16 02:00:00            Rogers Arena        away      16               CHI          3              4                   VAN          23       final         0         FALSE            1              1           0             1           0         0               0                 27                  4      0.8518519   4.102564  58.500000                    25                     3         0.8800000                     1                     1         0.0000000                     1                     0         1.0000000
#> 28:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030153 2011-04-18 00:00:00           United Center        home      16               CHI          2              3                   VAN          23       final         0         FALSE            1              1           0             1           0         0               0                 26                  3      0.8846154   3.078677  58.466667                    22                     2         0.9090909                     2                     1         0.5000000                     2                     0         1.0000000
#> 29:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030154 2011-04-20 00:00:00           United Center        home      16               CHI          7              2                   VAN          23       final         0         FALSE            1              1           1             0           0         0               0                 23                  2      0.9130435   2.017937  59.466667                    16                     0         1.0000000                     7                     2         0.7142857                     0                     0               NaN
#> 30:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030155 2011-04-22 02:00:00            Rogers Arena        away      16               CHI          5              0                   VAN          23       final         0         FALSE            1              1           1             0           0         0               1                 36                  0      1.0000000   0.000000  60.000000                    30                     0         1.0000000                     6                     0         1.0000000                     0                     0               NaN
#> 31:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030156 2011-04-24 23:30:00           United Center        home      16               CHI          4              3                   VAN          23       final         1         FALSE            1              1           1             0           0         0               0                 35                  3      0.9142857   2.384106  75.500000                    34                     3         0.9117647                     1                     0         1.0000000                     0                     0               NaN
#> 32:   8470645 Corey Crawford  20102011      2010-11    playoffs 2010030157 2011-04-27 02:00:00            Rogers Arena        away      16               CHI          1              2                   VAN          23       final         1         FALSE            1              1           0             1           0         0               0                 38                  2      0.9473684   1.838611  65.266667                    31                     2         0.9354839                     5                     0         1.0000000                     2                     0         1.0000000
#>     player_id    player_name season_id season_years season_type    game_id       game_datetime              venue_name team_status team_id team_abbreviation team_score opponent_score opponent_abbreviation opponent_id game_status game_nbot game_shootout goalie_games goalie_started goalie_wins goalie_losses goalie_ties goalie_ot goalie_shutouts goalie_shotagainst goalie_goalagainst goalie_savepct goalie_gaa goalie_toi goalie_ev_shotagainst goalie_ev_goalagainst goalie_ev_savepct goalie_pp_shotagainst goalie_pp_goalagainst goalie_pp_savepct goalie_pk_shotagainst goalie_pk_goalagainst goalie_pk_savepct