Get a tidy dataset of NHL players game logs

The function tidy_players_gamelogs() is meant to be a user-friendly way of getting the NHL game logs of selected players.

tidy_players_gamelogs(
  players_id,
  seasons_id = NULL,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = NULL
)

Arguments

players_id
Integer vector indicating the NHL ID of players for whom the game logs will be returned.

seasons_id
(optional) Character vector indicating the season for which the game logs will be returned. The required format for each character is 'xxxxyyyy'. Default to NULL which will return the full career game logs for each player.

regular
(optional) Logical indicating if the regular season schedule should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs schedule should be returned. Default to TRUE.

tz
(optional) Character specifying the timezone that should be used for datetime. Default to the user system timezone.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Load the data.table package to easily manipulate the data
library(data.table)

# Get the NHL career game logs of Vincent Lecavalier
gamelogs_lecavalier <- tidy_players_gamelogs(8467329L)

# Print the column names
colnames(gamelogs_lecavalier)

#>  [1] "player_name"           "season_years"          "season_type"          
#>  [4] "game_datetime"         "venue_name"            "team_status"          
#>  [7] "team_abbreviation"     "team_score"            "opponent_score"       
#> [10] "opponent_abbreviation" "game_status"           "game_nbot"            
#> [13] "game_shootout"         "skater_games"          "skater_goals"         
#> [16] "skater_assists"        "skater_points"         "skater_plusminus"     
#> [19] "skater_pim"            "skater_toi"            "skater_shifts"        
#> [22] "skater_gwg"            "skater_otg"            "skater_shots"         
#> [25] "skater_blocked"        "skater_hits"           "skater_ev_goals"      
#> [28] "skater_ev_assists"     "skater_ev_points"      "skater_ev_toi"        
#> [31] "skater_pp_goals"       "skater_pp_assists"     "skater_pp_points"     
#> [34] "skater_pp_toi"         "skater_pk_goals"       "skater_pk_assists"    
#> [37] "skater_pk_points"      "skater_pk_toi"         "goalie_games"         
#> [40] "goalie_started"        "goalie_wins"           "goalie_losses"        
#> [43] "goalie_ties"           "goalie_ot"             "goalie_shutouts"      
#> [46] "goalie_shotagainst"    "goalie_goalagainst"    "goalie_savepct"       
#> [49] "goalie_gaa"            "goalie_toi"            "goalie_ev_shotagainst"
#> [52] "goalie_ev_goalagainst" "goalie_ev_savepct"     "goalie_pp_shotagainst"
#> [55] "goalie_pp_goalagainst" "goalie_pp_savepct"     "goalie_pk_shotagainst"
#> [58] "goalie_pk_goalagainst" "goalie_pk_savepct"    

# Print an excerpt of the data
gamelogs_lecavalier[, .(season_years, season_type, skater_goals, skater_assists, skater_shots)]

#>       season_years season_type skater_goals skater_assists skater_shots
#>    1:      1998-99     regular            0              0            2
#>    2:      1998-99     regular            0              0            2
#>    3:      1998-99     regular            0              0            1
#>    4:      1998-99     regular            0              0            3
#>    5:      1998-99     regular            0              0            0
#>   ---                                                                  
#> 1283:      2015-16    playoffs            0              0            1
#> 1284:      2015-16    playoffs            1              0            1
#> 1285:      2015-16    playoffs            0              1            2
#> 1286:      2015-16    playoffs            0              0            0
#> 1287:      2015-16    playoffs            0              0            4

# Get the NHL 2010-11 playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
gamelogs_2011_playoffs_lecavalier_price <- tidy_players_gamelogs(
  players_id = c(8467329L, 8471679L),
  seasons_id = "20102011",
  regular    = FALSE,
  keep_id    = TRUE
)

# Print the column names
colnames(gamelogs_2011_playoffs_lecavalier_price)

#>  [1] "player_id"             "player_name"           "season_id"            
#>  [4] "season_years"          "season_type"           "game_id"              
#>  [7] "game_datetime"         "venue_name"            "team_status"          
#> [10] "team_id"               "team_abbreviation"     "team_score"           
#> [13] "opponent_score"        "opponent_abbreviation" "opponent_id"          
#> [16] "game_status"           "game_nbot"             "game_shootout"        
#> [19] "skater_games"          "skater_goals"          "skater_assists"       
#> [22] "skater_points"         "skater_plusminus"      "skater_pim"           
#> [25] "skater_toi"            "skater_shifts"         "skater_gwg"           
#> [28] "skater_otg"            "skater_shots"          "skater_blocked"       
#> [31] "skater_hits"           "skater_ev_goals"       "skater_ev_assists"    
#> [34] "skater_ev_points"      "skater_ev_toi"         "skater_pp_goals"      
#> [37] "skater_pp_assists"     "skater_pp_points"      "skater_pp_toi"        
#> [40] "skater_pk_goals"       "skater_pk_assists"     "skater_pk_points"     
#> [43] "skater_pk_toi"         "goalie_games"          "goalie_started"       
#> [46] "goalie_wins"           "goalie_losses"         "goalie_ties"          
#> [49] "goalie_ot"             "goalie_shutouts"       "goalie_shotagainst"   
#> [52] "goalie_goalagainst"    "goalie_savepct"        "goalie_gaa"           
#> [55] "goalie_toi"            "goalie_ev_shotagainst" "goalie_ev_goalagainst"
#> [58] "goalie_ev_savepct"     "goalie_pp_shotagainst" "goalie_pp_goalagainst"
#> [61] "goalie_pp_savepct"     "goalie_pk_shotagainst" "goalie_pk_goalagainst"
#> [64] "goalie_pk_savepct"    

# Print an excerpt of the data
gamelogs_2011_playoffs_lecavalier_price[, .(player_id, player_name, game_id, skater_points, goalie_savepct)]

#>     player_id        player_name    game_id skater_points goalie_savepct
#>  1:   8467329 Vincent Lecavalier 2010030141             0             NA
#>  2:   8467329 Vincent Lecavalier 2010030142             1             NA
#>  3:   8467329 Vincent Lecavalier 2010030143             1             NA
#>  4:   8467329 Vincent Lecavalier 2010030144             1             NA
#>  5:   8467329 Vincent Lecavalier 2010030145             2             NA
#>  6:   8467329 Vincent Lecavalier 2010030146             1             NA
#>  7:   8467329 Vincent Lecavalier 2010030147             0             NA
#>  8:   8467329 Vincent Lecavalier 2010030211             2             NA
#>  9:   8467329 Vincent Lecavalier 2010030212             2             NA
#> 10:   8467329 Vincent Lecavalier 2010030213             1             NA
#> 11:   8467329 Vincent Lecavalier 2010030214             1             NA
#> 12:   8467329 Vincent Lecavalier 2010030311             0             NA
#> 13:   8467329 Vincent Lecavalier 2010030312             4             NA
#> 14:   8467329 Vincent Lecavalier 2010030313             0             NA
#> 15:   8467329 Vincent Lecavalier 2010030314             1             NA
#> 16:   8467329 Vincent Lecavalier 2010030315             0             NA
#> 17:   8467329 Vincent Lecavalier 2010030316             2             NA
#> 18:   8467329 Vincent Lecavalier 2010030317             0             NA
#> 19:   8471679        Carey Price 2010030131            NA      1.0000000
#> 20:   8471679        Carey Price 2010030132            NA      0.9714286
#> 21:   8471679        Carey Price 2010030133            NA      0.8750000
#> 22:   8471679        Carey Price 2010030134            NA      0.8571429
#> 23:   8471679        Carey Price 2010030135            NA      0.9607843
#> 24:   8471679        Carey Price 2010030136            NA      0.9687500
#> 25:   8471679        Carey Price 2010030137            NA      0.8823529
#>     player_id        player_name    game_id skater_points goalie_savepct