Get a tidy dataset of NHL players statistics

The function tidy_players_stats() is meant to be a user-friendly way of getting the NHL year-by-year statistics of selected players.

tidy_players_stats(
  players_id,
  regular = TRUE,
  playoffs = TRUE,
  keep_id = FALSE,
  return_datatable = NULL
)

Arguments

players_id
Integer vector indicating the NHL ID of players for whom the statistics will be returned.

regular
(optional) Logical indicating if the regular season statistics should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs statistics should be returned. Default to TRUE.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Get the NHL career statistics of Vincent Lecavalier
tidy_players_stats(8467329L)

#>           player_name season_years season_type team_abbreviation skater_games
#> 1  Vincent Lecavalier      1998-99     regular               TBL           82
#> 2  Vincent Lecavalier      1999-00     regular               TBL           80
#> 3  Vincent Lecavalier      2000-01     regular               TBL           68
#> 4  Vincent Lecavalier      2001-02     regular               TBL           76
#> 5  Vincent Lecavalier      2002-03     regular               TBL           80
#> 6  Vincent Lecavalier      2002-03    playoffs               TBL           11
#> 7  Vincent Lecavalier      2003-04     regular               TBL           81
#> 8  Vincent Lecavalier      2003-04    playoffs               TBL           23
#> 9  Vincent Lecavalier      2005-06     regular               TBL           80
#> 10 Vincent Lecavalier      2005-06    playoffs               TBL            5
#> 11 Vincent Lecavalier      2006-07     regular               TBL           82
#> 12 Vincent Lecavalier      2006-07    playoffs               TBL            6
#> 13 Vincent Lecavalier      2007-08     regular               TBL           81
#> 14 Vincent Lecavalier      2008-09     regular               TBL           77
#> 15 Vincent Lecavalier      2009-10     regular               TBL           82
#> 16 Vincent Lecavalier      2010-11     regular               TBL           65
#> 17 Vincent Lecavalier      2010-11    playoffs               TBL           18
#> 18 Vincent Lecavalier      2011-12     regular               TBL           64
#> 19 Vincent Lecavalier      2012-13     regular               TBL           39
#> 20 Vincent Lecavalier      2013-14     regular               PHI           69
#> 21 Vincent Lecavalier      2013-14    playoffs               PHI            7
#> 22 Vincent Lecavalier      2014-15     regular               PHI           57
#> 23 Vincent Lecavalier      2015-16     regular               PHI            7
#> 24 Vincent Lecavalier      2015-16     regular               LAK           42
#> 25 Vincent Lecavalier      2015-16    playoffs               LAK            5
#>    skater_goals skater_assists skater_points skater_plusminus skater_pim
#> 1            13             15            28              -19         23
#> 2            25             42            67              -25         43
#> 3            23             28            51              -26         66
#> 4            20             17            37              -18         61
#> 5            33             45            78                0         39
#> 6             3              3             6               -2         22
#> 7            32             34            66               23         52
#> 8             9              7            16               -2         25
#> 9            35             40            75                0         90
#> 10            1              3             4                0          7
#> 11           52             56           108                2         44
#> 12            5              2             7                4         10
#> 13           40             52            92              -17         89
#> 14           29             38            67               -9         54
#> 15           24             46            70              -16         63
#> 16           25             29            54               -5         43
#> 17            6             13            19                6         16
#> 18           22             27            49               -2         50
#> 19           10             22            32               -5         29
#> 20           20             17            37              -16         44
#> 21            1              1             2               -5          2
#> 22            8             12            20               -7         36
#> 23            0              1             1               -1          2
#> 24           10              7            17                1         20
#> 25            1              1             2                0          2
#>    skater_toi skater_shifts skater_gwg skater_otg skater_shots skater_blocked
#> 1  1120.11667          1155          2          0          125              0
#> 2  1544.11667          1668          3          0          166              0
#> 3  1356.13333          1436          3          0          165              0
#> 4  1302.85000          1516          3          2          164              0
#> 5  1543.25000          1843          3          0          274             13
#> 6   248.53333           325          1          1           32              0
#> 7  1462.85000          1858          6          0          242              0
#> 8   451.83333           626          0          0           76              0
#> 9  1610.11667          1881          7          1          309             33
#> 10  111.43333           137          0          0           17              1
#> 11 1853.53333          2105          7          1          339             22
#> 12  158.86667           183          1          0           30              1
#> 13 1859.08333          2014          7          0          318             21
#> 14 1559.46667          1943          6          1          291             29
#> 15 1622.81667          2029          3          0          295             21
#> 16 1199.51667          1464          5          2          210             25
#> 17  357.50000           444          3          1           56              7
#> 18 1211.60000          1407          5          1          182             31
#> 19  697.43333           851          0          0           86             13
#> 20 1048.16667          1389          3          1          132             19
#> 21   74.80000           110          0          0            9              2
#> 22  720.90000           983          0          0          103             15
#> 23   66.21667            99          0          0            7              1
#> 24  582.55000           830          0          0           60             21
#> 25   71.66667           109          0          0            8              2
#>    skater_hits skater_ev_goals skater_ev_assists skater_ev_points skater_ev_toi
#> 1           52              11                10               21     918.91667
#> 2          116              19                20               39    1114.41667
#> 3           72              16                18               34     946.13333
#> 4           85              15                 8               23    1052.18333
#> 5           43              20                32               52    1134.86667
#> 6            8               2                 1                3     199.43333
#> 7           48              25                25               50    1136.56667
#> 8           33               7                 4               11     366.68333
#> 9           78              20                20               40    1100.68333
#> 10          10               0                 2                2      72.43333
#> 11         112              31                32               63    1340.31667
#> 12          21               4                 1                5     112.26667
#> 13         112              29                30               59    1394.66667
#> 14         110              18                23               41    1120.43333
#> 15          80              19                26               45    1234.46667
#> 16          93              13                13               26     925.20000
#> 17          20               3                 9               12     281.13333
#> 18         103              17                21               38     982.26667
#> 19          74               5                18               23     570.95000
#> 20          77              12                12               24     876.33333
#> 21           4               0                 0                0      61.61667
#> 22          40               7                 9               16     610.73333
#> 23           2               0                 1                1      63.78333
#> 24          60               4                 6               10     485.61667
#> 25          11               0                 1                1      62.43333
#>    skater_pp_goals skater_pp_assists skater_pp_points skater_pp_toi
#> 1                2                 4                6    189.733333
#> 2                6                22               28    417.566667
#> 3                7                10               17    389.300000
#> 4                5                 9               14    241.433333
#> 5               11                13               24    327.400000
#> 6                1                 2                3     35.700000
#> 7                5                 9               14    282.466667
#> 8                2                 3                5     78.800000
#> 9               13                19               32    414.483333
#> 10               1                 1                2     28.783333
#> 11              16                20               36    388.150000
#> 12               1                 1                2     32.816667
#> 13              10                19               29    337.950000
#> 14              10                14               24    327.883333
#> 15               5                20               25    331.416667
#> 16              12                16               28    259.266667
#> 17               3                 4                7     72.350000
#> 18               5                 6               11    217.633333
#> 19               5                 4                9    119.200000
#> 20               8                 5               13    169.250000
#> 21               1                 1                2     13.183333
#> 22               1                 3                4    109.983333
#> 23               0                 0                0      2.433333
#> 24               6                 1                7     73.266667
#> 25               1                 0                1      7.950000
#>    skater_pk_goals skater_pk_assists skater_pk_points skater_pk_toi
#> 1                0                 1                1    11.4666667
#> 2                0                 0                0    12.1333333
#> 3                0                 0                0    20.7000000
#> 4                0                 0                0     9.2333333
#> 5                2                 0                2    80.9833333
#> 6                0                 0                0    13.4000000
#> 7                2                 0                2    43.8166667
#> 8                0                 0                0     6.3500000
#> 9                2                 1                3    94.9500000
#> 10               0                 0                0    10.2166667
#> 11               5                 4                9   125.0666667
#> 12               0                 0                0    13.7833333
#> 13               1                 3                4   126.4666667
#> 14               1                 1                2   111.1500000
#> 15               0                 0                0    56.9333333
#> 16               0                 0                0    15.0500000
#> 17               0                 0                0     4.0166667
#> 18               0                 0                0    11.7000000
#> 19               0                 0                0     7.2833333
#> 20               0                 0                0     2.5833333
#> 21               0                 0                0     0.0000000
#> 22               0                 0                0     0.1833333
#> 23               0                 0                0     0.0000000
#> 24               0                 0                0    23.6666667
#> 25               0                 0                0     1.2833333
#>    goalies_games goalies_started goalie_wins goalie_losses goalie_ties
#> 1             NA              NA          NA            NA          NA
#> 2             NA              NA          NA            NA          NA
#> 3             NA              NA          NA            NA          NA
#> 4             NA              NA          NA            NA          NA
#> 5             NA              NA          NA            NA          NA
#> 6             NA              NA          NA            NA          NA
#> 7             NA              NA          NA            NA          NA
#> 8             NA              NA          NA            NA          NA
#> 9             NA              NA          NA            NA          NA
#> 10            NA              NA          NA            NA          NA
#> 11            NA              NA          NA            NA          NA
#> 12            NA              NA          NA            NA          NA
#> 13            NA              NA          NA            NA          NA
#> 14            NA              NA          NA            NA          NA
#> 15            NA              NA          NA            NA          NA
#> 16            NA              NA          NA            NA          NA
#> 17            NA              NA          NA            NA          NA
#> 18            NA              NA          NA            NA          NA
#> 19            NA              NA          NA            NA          NA
#> 20            NA              NA          NA            NA          NA
#> 21            NA              NA          NA            NA          NA
#> 22            NA              NA          NA            NA          NA
#> 23            NA              NA          NA            NA          NA
#> 24            NA              NA          NA            NA          NA
#> 25            NA              NA          NA            NA          NA
#>    goalie_ot goalie_shutouts goalie_shotagainst goalie_goalagainst
#> 1         NA              NA                 NA                 NA
#> 2         NA              NA                 NA                 NA
#> 3         NA              NA                 NA                 NA
#> 4         NA              NA                 NA                 NA
#> 5         NA              NA                 NA                 NA
#> 6         NA              NA                 NA                 NA
#> 7         NA              NA                 NA                 NA
#> 8         NA              NA                 NA                 NA
#> 9         NA              NA                 NA                 NA
#> 10        NA              NA                 NA                 NA
#> 11        NA              NA                 NA                 NA
#> 12        NA              NA                 NA                 NA
#> 13        NA              NA                 NA                 NA
#> 14        NA              NA                 NA                 NA
#> 15        NA              NA                 NA                 NA
#> 16        NA              NA                 NA                 NA
#> 17        NA              NA                 NA                 NA
#> 18        NA              NA                 NA                 NA
#> 19        NA              NA                 NA                 NA
#> 20        NA              NA                 NA                 NA
#> 21        NA              NA                 NA                 NA
#> 22        NA              NA                 NA                 NA
#> 23        NA              NA                 NA                 NA
#> 24        NA              NA                 NA                 NA
#> 25        NA              NA                 NA                 NA
#>    goalie_savepct goalie_gaa goalie_toi goalie_ev_shotagainst
#> 1              NA         NA         NA                    NA
#> 2              NA         NA         NA                    NA
#> 3              NA         NA         NA                    NA
#> 4              NA         NA         NA                    NA
#> 5              NA         NA         NA                    NA
#> 6              NA         NA         NA                    NA
#> 7              NA         NA         NA                    NA
#> 8              NA         NA         NA                    NA
#> 9              NA         NA         NA                    NA
#> 10             NA         NA         NA                    NA
#> 11             NA         NA         NA                    NA
#> 12             NA         NA         NA                    NA
#> 13             NA         NA         NA                    NA
#> 14             NA         NA         NA                    NA
#> 15             NA         NA         NA                    NA
#> 16             NA         NA         NA                    NA
#> 17             NA         NA         NA                    NA
#> 18             NA         NA         NA                    NA
#> 19             NA         NA         NA                    NA
#> 20             NA         NA         NA                    NA
#> 21             NA         NA         NA                    NA
#> 22             NA         NA         NA                    NA
#> 23             NA         NA         NA                    NA
#> 24             NA         NA         NA                    NA
#> 25             NA         NA         NA                    NA
#>    goalie_ev_goalagainst goalie_ev_savepct goalie_pp_shotagainst
#> 1                     NA                NA                    NA
#> 2                     NA                NA                    NA
#> 3                     NA                NA                    NA
#> 4                     NA                NA                    NA
#> 5                     NA                NA                    NA
#> 6                     NA                NA                    NA
#> 7                     NA                NA                    NA
#> 8                     NA                NA                    NA
#> 9                     NA                NA                    NA
#> 10                    NA                NA                    NA
#> 11                    NA                NA                    NA
#> 12                    NA                NA                    NA
#> 13                    NA                NA                    NA
#> 14                    NA                NA                    NA
#> 15                    NA                NA                    NA
#> 16                    NA                NA                    NA
#> 17                    NA                NA                    NA
#> 18                    NA                NA                    NA
#> 19                    NA                NA                    NA
#> 20                    NA                NA                    NA
#> 21                    NA                NA                    NA
#> 22                    NA                NA                    NA
#> 23                    NA                NA                    NA
#> 24                    NA                NA                    NA
#> 25                    NA                NA                    NA
#>    goalie_pp_goalagainst goalie_pp_savepct goalie_pk_shotagainst
#> 1                     NA                NA                    NA
#> 2                     NA                NA                    NA
#> 3                     NA                NA                    NA
#> 4                     NA                NA                    NA
#> 5                     NA                NA                    NA
#> 6                     NA                NA                    NA
#> 7                     NA                NA                    NA
#> 8                     NA                NA                    NA
#> 9                     NA                NA                    NA
#> 10                    NA                NA                    NA
#> 11                    NA                NA                    NA
#> 12                    NA                NA                    NA
#> 13                    NA                NA                    NA
#> 14                    NA                NA                    NA
#> 15                    NA                NA                    NA
#> 16                    NA                NA                    NA
#> 17                    NA                NA                    NA
#> 18                    NA                NA                    NA
#> 19                    NA                NA                    NA
#> 20                    NA                NA                    NA
#> 21                    NA                NA                    NA
#> 22                    NA                NA                    NA
#> 23                    NA                NA                    NA
#> 24                    NA                NA                    NA
#> 25                    NA                NA                    NA
#>    goalie_pk_goalagainst goalie_pk_savepct
#> 1                     NA                NA
#> 2                     NA                NA
#> 3                     NA                NA
#> 4                     NA                NA
#> 5                     NA                NA
#> 6                     NA                NA
#> 7                     NA                NA
#> 8                     NA                NA
#> 9                     NA                NA
#> 10                    NA                NA
#> 11                    NA                NA
#> 12                    NA                NA
#> 13                    NA                NA
#> 14                    NA                NA
#> 15                    NA                NA
#> 16                    NA                NA
#> 17                    NA                NA
#> 18                    NA                NA
#> 19                    NA                NA
#> 20                    NA                NA
#> 21                    NA                NA
#> 22                    NA                NA
#> 23                    NA                NA
#> 24                    NA                NA
#> 25                    NA                NA

# Get the NHL career playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
tidy_players_stats(
  players_id = c(8467329L, 8471679L),
  regular    = FALSE,
  keep_id    = TRUE
)

#>    player_id        player_name season_id season_years season_type team_id
#> 1    8467329 Vincent Lecavalier  20022003      2002-03    playoffs      14
#> 2    8467329 Vincent Lecavalier  20032004      2003-04    playoffs      14
#> 3    8467329 Vincent Lecavalier  20052006      2005-06    playoffs      14
#> 4    8467329 Vincent Lecavalier  20062007      2006-07    playoffs      14
#> 5    8467329 Vincent Lecavalier  20102011      2010-11    playoffs      14
#> 6    8467329 Vincent Lecavalier  20132014      2013-14    playoffs       4
#> 7    8467329 Vincent Lecavalier  20152016      2015-16    playoffs      26
#> 8    8471679        Carey Price  20072008      2007-08    playoffs       8
#> 9    8471679        Carey Price  20082009      2008-09    playoffs       8
#> 10   8471679        Carey Price  20092010      2009-10    playoffs       8
#> 11   8471679        Carey Price  20102011      2010-11    playoffs       8
#> 12   8471679        Carey Price  20122013      2012-13    playoffs       8
#> 13   8471679        Carey Price  20132014      2013-14    playoffs       8
#> 14   8471679        Carey Price  20142015      2014-15    playoffs       8
#> 15   8471679        Carey Price  20162017      2016-17    playoffs       8
#> 16   8471679        Carey Price  20192020      2019-20    playoffs       8
#>    team_abbreviation skater_games skater_goals skater_assists skater_points
#> 1                TBL           11            3              3             6
#> 2                TBL           23            9              7            16
#> 3                TBL            5            1              3             4
#> 4                TBL            6            5              2             7
#> 5                TBL           18            6             13            19
#> 6                PHI            7            1              1             2
#> 7                LAK            5            1              1             2
#> 8                MTL           NA           NA             NA            NA
#> 9                MTL           NA           NA             NA            NA
#> 10               MTL           NA           NA             NA            NA
#> 11               MTL           NA           NA             NA            NA
#> 12               MTL           NA           NA             NA            NA
#> 13               MTL           NA           NA             NA            NA
#> 14               MTL           NA           NA             NA            NA
#> 15               MTL           NA           NA             NA            NA
#> 16               MTL           NA           NA             NA            NA
#>    skater_plusminus skater_pim skater_toi skater_shifts skater_gwg skater_otg
#> 1                -2         22  248.53333           325          1          1
#> 2                -2         25  451.83333           626          0          0
#> 3                 0          7  111.43333           137          0          0
#> 4                 4         10  158.86667           183          1          0
#> 5                 6         16  357.50000           444          3          1
#> 6                -5          2   74.80000           110          0          0
#> 7                 0          2   71.66667           109          0          0
#> 8                NA         NA         NA            NA         NA         NA
#> 9                NA         NA         NA            NA         NA         NA
#> 10               NA         NA         NA            NA         NA         NA
#> 11               NA         NA         NA            NA         NA         NA
#> 12               NA         NA         NA            NA         NA         NA
#> 13               NA         NA         NA            NA         NA         NA
#> 14               NA         NA         NA            NA         NA         NA
#> 15               NA         NA         NA            NA         NA         NA
#> 16               NA         NA         NA            NA         NA         NA
#>    skater_shots skater_blocked skater_hits skater_ev_goals skater_ev_assists
#> 1            32              0           8               2                 1
#> 2            76              0          33               7                 4
#> 3            17              1          10               0                 2
#> 4            30              1          21               4                 1
#> 5            56              7          20               3                 9
#> 6             9              2           4               0                 0
#> 7             8              2          11               0                 1
#> 8            NA             NA          NA              NA                NA
#> 9            NA             NA          NA              NA                NA
#> 10           NA             NA          NA              NA                NA
#> 11           NA             NA          NA              NA                NA
#> 12           NA             NA          NA              NA                NA
#> 13           NA             NA          NA              NA                NA
#> 14           NA             NA          NA              NA                NA
#> 15           NA             NA          NA              NA                NA
#> 16           NA             NA          NA              NA                NA
#>    skater_ev_points skater_ev_toi skater_pp_goals skater_pp_assists
#> 1                 3     199.43333               1                 2
#> 2                11     366.68333               2                 3
#> 3                 2      72.43333               1                 1
#> 4                 5     112.26667               1                 1
#> 5                12     281.13333               3                 4
#> 6                 0      61.61667               1                 1
#> 7                 1      62.43333               1                 0
#> 8                NA            NA              NA                NA
#> 9                NA            NA              NA                NA
#> 10               NA            NA              NA                NA
#> 11               NA            NA              NA                NA
#> 12               NA            NA              NA                NA
#> 13               NA            NA              NA                NA
#> 14               NA            NA              NA                NA
#> 15               NA            NA              NA                NA
#> 16               NA            NA              NA                NA
#>    skater_pp_points skater_pp_toi skater_pk_goals skater_pk_assists
#> 1                 3      35.70000               0                 0
#> 2                 5      78.80000               0                 0
#> 3                 2      28.78333               0                 0
#> 4                 2      32.81667               0                 0
#> 5                 7      72.35000               0                 0
#> 6                 2      13.18333               0                 0
#> 7                 1       7.95000               0                 0
#> 8                NA            NA              NA                NA
#> 9                NA            NA              NA                NA
#> 10               NA            NA              NA                NA
#> 11               NA            NA              NA                NA
#> 12               NA            NA              NA                NA
#> 13               NA            NA              NA                NA
#> 14               NA            NA              NA                NA
#> 15               NA            NA              NA                NA
#> 16               NA            NA              NA                NA
#>    skater_pk_points skater_pk_toi goalies_games goalies_started goalie_wins
#> 1                 0     13.400000            NA              NA          NA
#> 2                 0      6.350000            NA              NA          NA
#> 3                 0     10.216667            NA              NA          NA
#> 4                 0     13.783333            NA              NA          NA
#> 5                 0      4.016667            NA              NA          NA
#> 6                 0      0.000000            NA              NA          NA
#> 7                 0      1.283333            NA              NA          NA
#> 8                NA            NA            11              11           5
#> 9                NA            NA             4               4           0
#> 10               NA            NA             4               1           0
#> 11               NA            NA             7               7           3
#> 12               NA            NA             4               4           1
#> 13               NA            NA            12              12           8
#> 14               NA            NA            12              12           6
#> 15               NA            NA             6               6           2
#> 16               NA            NA            10              10           5
#>    goalie_losses goalie_ties goalie_ot goalie_shutouts goalie_shotagainst
#> 1             NA          NA        NA              NA                 NA
#> 2             NA          NA        NA              NA                 NA
#> 3             NA          NA        NA              NA                 NA
#> 4             NA          NA        NA              NA                 NA
#> 5             NA          NA        NA              NA                 NA
#> 6             NA          NA        NA              NA                 NA
#> 7             NA          NA        NA              NA                 NA
#> 8              6          NA        NA               2                304
#> 9              4          NA        NA               0                123
#> 10             1          NA        NA               0                 73
#> 11             4          NA        NA               1                242
#> 12             2          NA        NA               0                123
#> 13             4          NA        NA               1                358
#> 14             6          NA        NA               1                352
#> 15             4          NA        NA               0                179
#> 16             5          NA         0               2                282
#>    goalie_goalagainst goalie_savepct goalie_gaa goalie_toi
#> 1                  NA             NA         NA         NA
#> 2                  NA             NA         NA         NA
#> 3                  NA             NA         NA         NA
#> 4                  NA             NA         NA         NA
#> 5                  NA             NA         NA         NA
#> 6                  NA             NA         NA         NA
#> 7                  NA             NA         NA         NA
#> 8                  30      0.9013158   2.776136   648.3833
#> 9                  15      0.8780488   4.106464   219.1667
#> 10                  8      0.8904110   3.564356   134.6667
#> 11                 16      0.9338843   2.107651   455.4833
#> 12                 13      0.8943089   3.262461   239.0833
#> 13                 29      0.9189944   2.355277   738.7667
#> 14                 28      0.9204545   2.234885   751.7167
#> 15                 12      0.9329609   1.858224   387.4667
#> 16                 18      0.9361702   1.783552   605.5333
#>    goalie_ev_shotagainst goalie_ev_goalagainst goalie_ev_savepct
#> 1                     NA                    NA                NA
#> 2                     NA                    NA                NA
#> 3                     NA                    NA                NA
#> 4                     NA                    NA                NA
#> 5                     NA                    NA                NA
#> 6                     NA                    NA                NA
#> 7                     NA                    NA                NA
#> 8                    232                    23         0.9008621
#> 9                     99                    11         0.8888889
#> 10                    57                     6         0.8947368
#> 11                   216                    16         0.9259259
#> 12                   101                    10         0.9009901
#> 13                   309                    23         0.9255663
#> 14                   288                    16         0.9444444
#> 15                   156                    10         0.9358974
#> 16                   208                    11         0.9471154
#>    goalie_pp_shotagainst goalie_pp_goalagainst goalie_pp_savepct
#> 1                     NA                    NA                NA
#> 2                     NA                    NA                NA
#> 3                     NA                    NA                NA
#> 4                     NA                    NA                NA
#> 5                     NA                    NA                NA
#> 6                     NA                    NA                NA
#> 7                     NA                    NA                NA
#> 8                     53                     5         0.9056604
#> 9                     22                     4         0.8181818
#> 10                    15                     1         0.9333333
#> 11                    21                     0         1.0000000
#> 12                    20                     3         0.8500000
#> 13                    42                     5         0.8809524
#> 14                    51                    12         0.7647059
#> 15                    18                     1         0.9444444
#> 16                    70                     7         0.9000000
#>    goalie_pk_shotagainst goalie_pk_goalagainst goalie_pk_savepct
#> 1                     NA                    NA                NA
#> 2                     NA                    NA                NA
#> 3                     NA                    NA                NA
#> 4                     NA                    NA                NA
#> 5                     NA                    NA                NA
#> 6                     NA                    NA                NA
#> 7                     NA                    NA                NA
#> 8                     19                     2         0.8947368
#> 9                      2                     0         1.0000000
#> 10                     1                     1         0.0000000
#> 11                     5                     0         1.0000000
#> 12                     2                     0         1.0000000
#> 13                     7                     1         0.8571429
#> 14                    13                     0         1.0000000
#> 15                     5                     1         0.8000000
#> 16                     4                     0         1.0000000