Get a tidy dataset of the NHL schedule

The function tidy_schedule() is meant to be a user-friendly way of getting the NHL schedule, including the final score of completed games.

tidy_schedule(
  seasons_id,
  season = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = NULL
)

Arguments

seasons_id
Character vector of the seasons ID for which the schedule will be returned. The required format is 'xxxxyyyy'.

season
(optional) Logical indicating if the season schedule should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs schedule should be returned. Default to TRUE.

tz
(optional) Character specifying the timezone that should be used for datetime. Default to the user system timezone.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Get the schedule of the 2019-2020 season and playoffs
head(tidy_schedule("20192020"))

#>   game_type       game_datetime game_status            venue_name away_team
#> 1         R 2019-10-02 19:00:00       Final      Scotiabank Arena       OTT
#> 2         R 2019-10-02 20:00:00       Final     Enterprise Center       WSH
#> 3         R 2019-10-02 22:00:00       Final          Rogers Place       VAN
#> 4         R 2019-10-02 22:30:00       Final        T-Mobile Arena       SJS
#> 5         R 2019-10-03 19:00:00       Final          Amalie Arena       FLA
#> 6         R 2019-10-03 19:00:00       Final Madison Square Garden       WPG
#>   away_score home_score home_team game_nbot game_shootout
#> 1          3          5       TOR         0         FALSE
#> 2          3          2       STL         1         FALSE
#> 3          2          3       EDM         0         FALSE
#> 4          1          4       VGK         0         FALSE
#> 5          2          5       TBL         0         FALSE
#> 6          4          6       NYR         0         FALSE

# Get the regular season schedule of both the 2018-2019 and 2019-2020 seasons,
# keeping the IDs and indicating game datetime with Los Angeles local time
head(tidy_schedule(
  seasons_id = c("20182019", "20192020"),
  playoffs   = FALSE,
  tz         = "America/Los_Angeles",
  keep_id    = TRUE
))

#>   season_id    game_id game_type       game_datetime game_status
#> 1  20182019 2018020001         R 2018-10-03 16:00:00       Final
#> 2  20182019 2018020002         R 2018-10-03 16:30:00       Final
#> 3  20182019 2018020003         R 2018-10-03 19:00:00       Final
#> 4  20182019 2018020004         R 2018-10-03 19:30:00       Final
#> 5  20182019 2018020005         R 2018-10-04 16:00:00       Final
#> 6  20182019 2018020007         R 2018-10-04 16:00:00       Final
#>               venue_name away_id away_team away_score home_score home_team
#> 1       Scotiabank Arena       8       MTL          2          3       TOR
#> 2      Capital One Arena       6       BOS          0          7       WSH
#> 3           Rogers Arena      20       CGY          2          5       VAN
#> 4 SAP Center at San Jose      24       ANA          5          2       SJS
#> 5         KeyBank Center       6       BOS          4          0       BUF
#> 6       PPG Paints Arena      15       WSH          6          7       PIT
#>   home_id game_nbot game_shootout
#> 1      10         1         FALSE
#> 2      15         0         FALSE
#> 3      23         0         FALSE
#> 4      28         0         FALSE
#> 5       7         0         FALSE
#> 6       5         1         FALSE