Get a tidy dataset of NHL teams metadata

The function tidy_teams_meta() is meant to be a user-friendly way of getting metadata about NHL teams.

tidy_teams_meta(
  active_only = TRUE,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)

Arguments

active_only
(optional) Logical indicating if only active teams should be returned. Default to TRUE.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default can be set globally with options("tidynhl.data.table").

Examples

# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get metadata of active teams
tidy_teams_meta()

#>     team_abbreviation   team_place      team_name         team_fullname team_shortname season_first_years conference_active_abbreviation conference_active_name division_active_abbreviation division_active_name        venue_active_name venue_active_country venue_active_stateprovince venue_active_city     venue_active_tz venue_active_lat venue_active_long                                       logo_last_url
#>  1:               ANA      Anaheim          Ducks         Anaheim Ducks        Anaheim            1993-94                              W                Western                          WST           Honda West             Honda Center                  USA                         CA           Anaheim America/Los_Angeles         33.80778        -117.87667 https://assets.nhle.com/logos/nhl/svg/ANA_light.svg
#>  2:               ARI      Arizona        Coyotes       Arizona Coyotes        Arizona            2014-15                              W                Western                          WST           Honda West         Gila River Arena                  USA                         AZ          Glendale     America/Phoenix         33.53194        -112.26111 https://assets.nhle.com/logos/nhl/svg/ARI_light.svg
#>  3:               BOS       Boston         Bruins         Boston Bruins         Boston            1924-25                              E                Eastern                          EST      MassMutual East                TD Garden                  USA                         MA            Boston    America/New_York         42.36630         -71.06223 https://assets.nhle.com/logos/nhl/svg/BOS_light.svg
#>  4:               BUF      Buffalo         Sabres        Buffalo Sabres        Buffalo            1970-71                              E                Eastern                          EST      MassMutual East           KeyBank Center                  USA                         NY           Buffalo    America/New_York         42.87500         -78.87639 https://assets.nhle.com/logos/nhl/svg/BUF_light.svg
#>  5:               CAR     Carolina     Hurricanes   Carolina Hurricanes       Carolina            1997-98                              E                Eastern                          CEN     Discover Central                PNC Arena                  USA                         NC           Raleigh    America/New_York         35.80333         -78.72194 https://assets.nhle.com/logos/nhl/svg/CAR_light.svg
#>  6:               CBJ     Columbus   Blue Jackets Columbus Blue Jackets       Columbus            2000-01                              E                Eastern                          CEN     Discover Central         Nationwide Arena                  USA                         OH          Columbus    America/New_York         39.96928         -83.00611 https://assets.nhle.com/logos/nhl/svg/CBJ_light.svg
#>  7:               CGY      Calgary         Flames        Calgary Flames        Calgary            1980-81                              W                Western                          NTH         Scotia North    Scotiabank Saddledome               Canada                         AB           Calgary      America/Denver         51.03750        -114.05194 https://assets.nhle.com/logos/nhl/svg/CGY_light.svg
#>  8:               CHI      Chicago     Blackhawks    Chicago Blackhawks        Chicago            1926-27                              W                Western                          CEN     Discover Central            United Center                  USA                         IL           Chicago     America/Chicago         41.88056         -87.67417 https://assets.nhle.com/logos/nhl/svg/CHI_light.svg
#>  9:               COL     Colorado      Avalanche    Colorado Avalanche       Colorado            1995-96                              W                Western                          WST           Honda West               Ball Arena                  USA                         CO            Denver      America/Denver         39.74861        -105.00750 https://assets.nhle.com/logos/nhl/svg/COL_light.svg
#> 10:               DAL       Dallas          Stars          Dallas Stars         Dallas            1993-94                              W                Western                          CEN     Discover Central American Airlines Center                  USA                         TX            Dallas     America/Chicago         32.79056         -96.81028 https://assets.nhle.com/logos/nhl/svg/DAL_light.svg
#> 11:               DET      Detroit      Red Wings     Detroit Red Wings        Detroit            1932-33                              E                Eastern                          CEN     Discover Central     Little Caesars Arena                  USA                         MI           Detroit     America/Detroit         42.34111         -83.05500 https://assets.nhle.com/logos/nhl/svg/DET_light.svg
#> 12:               EDM     Edmonton         Oilers       Edmonton Oilers       Edmonton            1979-80                              W                Western                          NTH         Scotia North             Rogers Place               Canada                         AB          Edmonton    America/Edmonton         53.54694        -113.49778 https://assets.nhle.com/logos/nhl/svg/EDM_light.svg
#> 13:               FLA      Florida       Panthers      Florida Panthers        Florida            1993-94                              E                Eastern                          CEN     Discover Central              BB&T Center                  USA                         FL           Sunrise    America/New_York         26.15833         -80.32556 https://assets.nhle.com/logos/nhl/svg/FLA_light.svg
#> 14:               LAK  Los Angeles          Kings     Los Angeles Kings    Los Angeles            1967-68                              W                Western                          WST           Honda West           STAPLES Center                  USA                         CA       Los Angeles America/Los_Angeles         34.04306        -118.26722 https://assets.nhle.com/logos/nhl/svg/LAK_light.svg
#> 15:               MIN    Minnesota           Wild        Minnesota Wild      Minnesota            2000-01                              W                Western                          WST           Honda West       Xcel Energy Center                  USA                         MN           St Paul     America/Chicago         44.94472         -93.10111 https://assets.nhle.com/logos/nhl/svg/MIN_light.svg
#> 16:               MTL     Montréal      Canadiens    Montréal Canadiens       Montréal            1917-18                              E                Eastern                          NTH         Scotia North              Bell Centre               Canada                         QC          Montreal    America/Montreal         45.49611         -73.56944 https://assets.nhle.com/logos/nhl/svg/MTL_light.svg
#> 17:               NJD   New Jersey         Devils     New Jersey Devils     New Jersey            1982-83                              E                Eastern                          EST      MassMutual East        Prudential Center                  USA                         NJ            Newark    America/New_York         40.73361         -74.17111 https://assets.nhle.com/logos/nhl/svg/NJD_light.svg
#> 18:               NSH    Nashville      Predators   Nashville Predators      Nashville            1998-99                              W                Western                          CEN     Discover Central        Bridgestone Arena                  USA                         TN         Nashville     America/Chicago         36.15917         -86.77861 https://assets.nhle.com/logos/nhl/svg/NSH_light.svg
#> 19:               NYI     New York      Islanders    New York Islanders   NY Islanders            1972-73                              E                Eastern                          EST      MassMutual East          Nassau Coliseum                  USA                         NY         Uniondale    America/New_York         40.72278         -73.59056 https://assets.nhle.com/logos/nhl/svg/NYI_light.svg
#> 20:               NYR     New York        Rangers      New York Rangers     NY Rangers            1926-27                              E                Eastern                          EST      MassMutual East    Madison Square Garden                  USA                         NY          New York    America/New_York         40.75056         -73.99361 https://assets.nhle.com/logos/nhl/svg/NYR_light.svg
#> 21:               OTT       Ottawa       Senators       Ottawa Senators         Ottawa            1992-93                              E                Eastern                          NTH         Scotia North     Canadian Tire Centre               Canada                         ON            Ottawa    America/New_York         45.29694         -75.92722 https://assets.nhle.com/logos/nhl/svg/OTT_light.svg
#> 22:               PHI Philadelphia         Flyers   Philadelphia Flyers   Philadelphia            1967-68                              E                Eastern                          EST      MassMutual East       Wells Fargo Center                  USA                         PA      Philadelphia    America/New_York         39.90111         -75.17194 https://assets.nhle.com/logos/nhl/svg/PHI_light.svg
#> 23:               PIT   Pittsburgh       Penguins   Pittsburgh Penguins     Pittsburgh            1967-68                              E                Eastern                          EST      MassMutual East         PPG Paints Arena                  USA                         PA        Pittsburgh    America/New_York         40.43944         -79.98917 https://assets.nhle.com/logos/nhl/svg/PIT_light.svg
#> 24:               SJS     San Jose         Sharks       San Jose Sharks       San Jose            1991-92                              W                Western                          WST           Honda West   SAP Center at San Jose                  USA                         CA          San Jose America/Los_Angeles         37.33278        -121.90111 https://assets.nhle.com/logos/nhl/svg/SJS_light.svg
#> 25:               STL    St. Louis          Blues       St. Louis Blues       St Louis            1967-68                              W                Western                          WST           Honda West        Enterprise Center                  USA                         MO         St. Louis     America/Chicago         38.62667         -90.20250 https://assets.nhle.com/logos/nhl/svg/STL_light.svg
#> 26:               TBL    Tampa Bay      Lightning   Tampa Bay Lightning      Tampa Bay            1992-93                              E                Eastern                          CEN     Discover Central             AMALIE Arena                  USA                         FL             Tampa    America/New_York         27.94278         -82.45194 https://assets.nhle.com/logos/nhl/svg/TBL_light.svg
#> 27:               TOR      Toronto    Maple Leafs   Toronto Maple Leafs        Toronto            1926-27                              E                Eastern                          NTH         Scotia North         Scotiabank Arena               Canada                         ON           Toronto     America/Toronto         43.64333         -79.37917 https://assets.nhle.com/logos/nhl/svg/TOR_light.svg
#> 28:               VAN    Vancouver        Canucks     Vancouver Canucks      Vancouver            1970-71                              W                Western                          NTH         Scotia North             Rogers Arena               Canada                         BC         Vancouver   America/Vancouver         49.27778        -123.10889 https://assets.nhle.com/logos/nhl/svg/VAN_light.svg
#> 29:               VGK        Vegas Golden Knights  Vegas Golden Knights          Vegas            2017-18                              W                Western                          WST           Honda West           T-Mobile Arena                  USA                         NV         Las Vegas America/Los_Angeles         36.10278        -115.17833 https://assets.nhle.com/logos/nhl/svg/VGK_light.svg
#> 30:               WPG     Winnipeg           Jets         Winnipeg Jets       Winnipeg            2011-12                              W                Western                          NTH         Scotia North           Bell MTS Place               Canada                         MB          Winnipeg    America/Winnipeg         49.89278         -97.14361 https://assets.nhle.com/logos/nhl/svg/WPG_light.svg
#> 31:               WSH   Washington       Capitals   Washington Capitals     Washington            1974-75                              E                Eastern                          EST      MassMutual East        Capital One Arena                  USA                         DC        Washington    America/New_York         38.89806         -77.02083 https://assets.nhle.com/logos/nhl/svg/WSH_light.svg
#>     team_abbreviation   team_place      team_name         team_fullname team_shortname season_first_years conference_active_abbreviation conference_active_name division_active_abbreviation division_active_name        venue_active_name venue_active_country venue_active_stateprovince venue_active_city     venue_active_tz venue_active_lat venue_active_long                                       logo_last_url

# Get metadata of active and historical teams keeping the IDs
tidy_teams_meta(active_only = FALSE, keep_id = TRUE)

#>     franchise_id team_id team_abbreviation team_active   team_place       team_name           team_fullname team_shortname season_first_id season_first_years season_last_id season_last_years conference_active_id conference_active_abbreviation conference_active_name division_active_id division_active_abbreviation division_active_name        venue_active_name venue_active_country venue_active_stateprovince venue_active_city     venue_active_tz venue_active_lat venue_active_long                                                         logo_last_url
#>  1:           32      24               ANA        TRUE      Anaheim           Ducks           Anaheim Ducks        Anaheim        19931994            1993-94           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West             Honda Center                  USA                         CA           Anaheim America/Los_Angeles         33.80778        -117.87667                   https://assets.nhle.com/logos/nhl/svg/ANA_light.svg
#>  2:           28      53               ARI        TRUE      Arizona         Coyotes         Arizona Coyotes        Arizona        20142015            2014-15           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West         Gila River Arena                  USA                         AZ          Glendale     America/Phoenix         33.53194        -112.26111                   https://assets.nhle.com/logos/nhl/svg/ARI_light.svg
#>  3:            6       6               BOS        TRUE       Boston          Bruins           Boston Bruins         Boston        19241925            1924-25           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East                TD Garden                  USA                         MA            Boston    America/New_York         42.36630         -71.06223                   https://assets.nhle.com/logos/nhl/svg/BOS_light.svg
#>  4:           19       7               BUF        TRUE      Buffalo          Sabres          Buffalo Sabres        Buffalo        19701971            1970-71           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East           KeyBank Center                  USA                         NY           Buffalo    America/New_York         42.87500         -78.87639                   https://assets.nhle.com/logos/nhl/svg/BUF_light.svg
#>  5:           26      12               CAR        TRUE     Carolina      Hurricanes     Carolina Hurricanes       Carolina        19971998            1997-98           <NA>              <NA>                    6                              E                Eastern                 26                          CEN     Discover Central                PNC Arena                  USA                         NC           Raleigh    America/New_York         35.80333         -78.72194                   https://assets.nhle.com/logos/nhl/svg/CAR_light.svg
#>  6:           36      29               CBJ        TRUE     Columbus    Blue Jackets   Columbus Blue Jackets       Columbus        20002001            2000-01           <NA>              <NA>                    6                              E                Eastern                 26                          CEN     Discover Central         Nationwide Arena                  USA                         OH          Columbus    America/New_York         39.96928         -83.00611                   https://assets.nhle.com/logos/nhl/svg/CBJ_light.svg
#>  7:           21      20               CGY        TRUE      Calgary          Flames          Calgary Flames        Calgary        19801981            1980-81           <NA>              <NA>                    5                              W                Western                 28                          NTH         Scotia North    Scotiabank Saddledome               Canada                         AB           Calgary      America/Denver         51.03750        -114.05194                   https://assets.nhle.com/logos/nhl/svg/CGY_light.svg
#>  8:           11      16               CHI        TRUE      Chicago      Blackhawks      Chicago Blackhawks        Chicago        19261927            1926-27           <NA>              <NA>                    5                              W                Western                 26                          CEN     Discover Central            United Center                  USA                         IL           Chicago     America/Chicago         41.88056         -87.67417                   https://assets.nhle.com/logos/nhl/svg/CHI_light.svg
#>  9:           27      21               COL        TRUE     Colorado       Avalanche      Colorado Avalanche       Colorado        19951996            1995-96           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West               Ball Arena                  USA                         CO            Denver      America/Denver         39.74861        -105.00750                   https://assets.nhle.com/logos/nhl/svg/COL_light.svg
#> 10:           15      25               DAL        TRUE       Dallas           Stars            Dallas Stars         Dallas        19931994            1993-94           <NA>              <NA>                    5                              W                Western                 26                          CEN     Discover Central American Airlines Center                  USA                         TX            Dallas     America/Chicago         32.79056         -96.81028                   https://assets.nhle.com/logos/nhl/svg/DAL_light.svg
#> 11:           12      17               DET        TRUE      Detroit       Red Wings       Detroit Red Wings        Detroit        19321933            1932-33           <NA>              <NA>                    6                              E                Eastern                 26                          CEN     Discover Central     Little Caesars Arena                  USA                         MI           Detroit     America/Detroit         42.34111         -83.05500                   https://assets.nhle.com/logos/nhl/svg/DET_light.svg
#> 12:           25      22               EDM        TRUE     Edmonton          Oilers         Edmonton Oilers       Edmonton        19791980            1979-80           <NA>              <NA>                    5                              W                Western                 28                          NTH         Scotia North             Rogers Place               Canada                         AB          Edmonton    America/Edmonton         53.54694        -113.49778                   https://assets.nhle.com/logos/nhl/svg/EDM_light.svg
#> 13:           33      13               FLA        TRUE      Florida        Panthers        Florida Panthers        Florida        19931994            1993-94           <NA>              <NA>                    6                              E                Eastern                 26                          CEN     Discover Central              BB&T Center                  USA                         FL           Sunrise    America/New_York         26.15833         -80.32556                   https://assets.nhle.com/logos/nhl/svg/FLA_light.svg
#> 14:           14      26               LAK        TRUE  Los Angeles           Kings       Los Angeles Kings    Los Angeles        19671968            1967-68           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West           STAPLES Center                  USA                         CA       Los Angeles America/Los_Angeles         34.04306        -118.26722                   https://assets.nhle.com/logos/nhl/svg/LAK_light.svg
#> 15:           37      30               MIN        TRUE    Minnesota            Wild          Minnesota Wild      Minnesota        20002001            2000-01           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West       Xcel Energy Center                  USA                         MN           St Paul     America/Chicago         44.94472         -93.10111                   https://assets.nhle.com/logos/nhl/svg/MIN_light.svg
#> 16:            1       8               MTL        TRUE     Montréal       Canadiens      Montréal Canadiens       Montréal        19171918            1917-18           <NA>              <NA>                    6                              E                Eastern                 28                          NTH         Scotia North              Bell Centre               Canada                         QC          Montreal    America/Montreal         45.49611         -73.56944                   https://assets.nhle.com/logos/nhl/svg/MTL_light.svg
#> 17:           23       1               NJD        TRUE   New Jersey          Devils       New Jersey Devils     New Jersey        19821983            1982-83           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East        Prudential Center                  USA                         NJ            Newark    America/New_York         40.73361         -74.17111                   https://assets.nhle.com/logos/nhl/svg/NJD_light.svg
#> 18:           34      18               NSH        TRUE    Nashville       Predators     Nashville Predators      Nashville        19981999            1998-99           <NA>              <NA>                    5                              W                Western                 26                          CEN     Discover Central        Bridgestone Arena                  USA                         TN         Nashville     America/Chicago         36.15917         -86.77861                   https://assets.nhle.com/logos/nhl/svg/NSH_light.svg
#> 19:           22       2               NYI        TRUE     New York       Islanders      New York Islanders   NY Islanders        19721973            1972-73           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East          Nassau Coliseum                  USA                         NY         Uniondale    America/New_York         40.72278         -73.59056                   https://assets.nhle.com/logos/nhl/svg/NYI_light.svg
#> 20:           10       3               NYR        TRUE     New York         Rangers        New York Rangers     NY Rangers        19261927            1926-27           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East    Madison Square Garden                  USA                         NY          New York    America/New_York         40.75056         -73.99361                   https://assets.nhle.com/logos/nhl/svg/NYR_light.svg
#> 21:           30       9               OTT        TRUE       Ottawa        Senators         Ottawa Senators         Ottawa        19921993            1992-93           <NA>              <NA>                    6                              E                Eastern                 28                          NTH         Scotia North     Canadian Tire Centre               Canada                         ON            Ottawa    America/New_York         45.29694         -75.92722                   https://assets.nhle.com/logos/nhl/svg/OTT_light.svg
#> 22:           16       4               PHI        TRUE Philadelphia          Flyers     Philadelphia Flyers   Philadelphia        19671968            1967-68           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East       Wells Fargo Center                  USA                         PA      Philadelphia    America/New_York         39.90111         -75.17194                   https://assets.nhle.com/logos/nhl/svg/PHI_light.svg
#> 23:           17       5               PIT        TRUE   Pittsburgh        Penguins     Pittsburgh Penguins     Pittsburgh        19671968            1967-68           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East         PPG Paints Arena                  USA                         PA        Pittsburgh    America/New_York         40.43944         -79.98917                   https://assets.nhle.com/logos/nhl/svg/PIT_light.svg
#> 24:           29      28               SJS        TRUE     San Jose          Sharks         San Jose Sharks       San Jose        19911992            1991-92           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West   SAP Center at San Jose                  USA                         CA          San Jose America/Los_Angeles         37.33278        -121.90111                   https://assets.nhle.com/logos/nhl/svg/SJS_light.svg
#> 25:           18      19               STL        TRUE    St. Louis           Blues         St. Louis Blues       St Louis        19671968            1967-68           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West        Enterprise Center                  USA                         MO         St. Louis     America/Chicago         38.62667         -90.20250                   https://assets.nhle.com/logos/nhl/svg/STL_light.svg
#> 26:           31      14               TBL        TRUE    Tampa Bay       Lightning     Tampa Bay Lightning      Tampa Bay        19921993            1992-93           <NA>              <NA>                    6                              E                Eastern                 26                          CEN     Discover Central             AMALIE Arena                  USA                         FL             Tampa    America/New_York         27.94278         -82.45194                   https://assets.nhle.com/logos/nhl/svg/TBL_light.svg
#> 27:            5      10               TOR        TRUE      Toronto     Maple Leafs     Toronto Maple Leafs        Toronto        19261927            1926-27           <NA>              <NA>                    6                              E                Eastern                 28                          NTH         Scotia North         Scotiabank Arena               Canada                         ON           Toronto     America/Toronto         43.64333         -79.37917                   https://assets.nhle.com/logos/nhl/svg/TOR_light.svg
#> 28:           20      23               VAN        TRUE    Vancouver         Canucks       Vancouver Canucks      Vancouver        19701971            1970-71           <NA>              <NA>                    5                              W                Western                 28                          NTH         Scotia North             Rogers Arena               Canada                         BC         Vancouver   America/Vancouver         49.27778        -123.10889                   https://assets.nhle.com/logos/nhl/svg/VAN_light.svg
#> 29:           38      54               VGK        TRUE        Vegas  Golden Knights    Vegas Golden Knights          Vegas        20172018            2017-18           <NA>              <NA>                    5                              W                Western                 27                          WST           Honda West           T-Mobile Arena                  USA                         NV         Las Vegas America/Los_Angeles         36.10278        -115.17833                   https://assets.nhle.com/logos/nhl/svg/VGK_light.svg
#> 30:           35      52               WPG        TRUE     Winnipeg            Jets           Winnipeg Jets       Winnipeg        20112012            2011-12           <NA>              <NA>                    5                              W                Western                 28                          NTH         Scotia North           Bell MTS Place               Canada                         MB          Winnipeg    America/Winnipeg         49.89278         -97.14361                   https://assets.nhle.com/logos/nhl/svg/WPG_light.svg
#> 31:           24      15               WSH        TRUE   Washington        Capitals     Washington Capitals     Washington        19741975            1974-75           <NA>              <NA>                    6                              E                Eastern                 25                          EST      MassMutual East        Capital One Arena                  USA                         DC        Washington    America/New_York         38.89806         -77.02083                   https://assets.nhle.com/logos/nhl/svg/WSH_light.svg
#> 32:            2      41               MWN       FALSE     Montreal       Wanderers      Montreal Wanderers       Montreal        19171918            1917-18       19171918           1917-18                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/MWN_19171918_light.svg
#> 33:            5      57               TAN       FALSE      Toronto          Arenas          Toronto Arenas        Toronto        19171918            1917-18       19181919           1918-19                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/TAN_19171918-19181919_light.svg
#> 34:            4      42               QBD       FALSE       Quebec        Bulldogs         Quebec Bulldogs         Quebec        19191920            1919-20       19191920           1919-20                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/QBD_19191920_light.svg
#> 35:            4      37               HAM       FALSE     Hamilton          Tigers         Hamilton Tigers       Hamilton        19201921            1920-21       19241925           1924-25                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/HAM_19231924-19241925_light.svg
#> 36:            5      58               TSP       FALSE      Toronto    St. Patricks    Toronto St. Patricks        Toronto        19191920            1919-20       19251926           1925-26                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/TSP_19261927_light.svg
#> 37:           12      40               DCG       FALSE      Detroit         Cougars         Detroit Cougars        Detroit        19261927            1926-27       19291930           1929-30                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/DCG_19261927-19291930_light.svg
#> 38:            9      38               PIR       FALSE   Pittsburgh         Pirates      Pittsburgh Pirates     Pittsburgh        19251926            1925-26       19291930           1929-30                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/PIR_19291930_light.svg
#> 39:            9      39               QUA       FALSE Philadelphia         Quakers    Philadelphia Quakers   Philadelphia        19301931            1930-31       19301931           1930-31                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/QUA_19301931_light.svg
#> 40:           12      50               DFL       FALSE      Detroit         Falcons         Detroit Falcons        Detroit        19301931            1930-31       19311932           1931-32                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/DFL_19301931-19311932_light.svg
#> 41:            3      36               SEN       FALSE       Ottawa Senators (1917)  Ottawa Senators (1917)         Ottawa        19171918            1917-18       19331934           1933-34                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/SEN_19171918-19331934_light.svg
#> 42:            3      45               SLE       FALSE    St. Louis          Eagles        St. Louis Eagles      St. Louis        19341935            1934-35       19341935           1934-35                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/SLE_19341935_light.svg
#> 43:            7      43               MMR       FALSE     Montreal         Maroons        Montreal Maroons       Montreal        19241925            1924-25       19371938           1937-38                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/MMR_19351936-19371938_light.svg
#> 44:            8      44               NYA       FALSE     New York       Americans      New York Americans       New York        19251926            1925-26       19401941           1940-41                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/NYA_19401941_light.svg
#> 45:            8      51               BRK       FALSE     Brooklyn       Americans      Brooklyn Americans       Brooklyn        19411942            1941-42       19411942           1941-42                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA          https://assets.nhle.com/logos/nhl/svg/BRK_19411942_light.svg
#> 46:           13      46               OAK       FALSE      Oakland           Seals           Oakland Seals        Oakland        19671968            1967-68       19691970           1969-70                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/OAK_19671968-19691970_light.svg
#> 47:           13      56               CGS       FALSE   California    Golden Seals California Golden Seals     California        19701971            1970-71       19751976           1975-76                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/CGS_19741975-19751976_light.svg
#> 48:           23      48               KCS       FALSE  Kansas City          Scouts      Kansas City Scouts    Kansas City        19741975            1974-75       19751976           1975-76                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/KCS_19741975-19751976_light.svg
#> 49:           13      49               CLE       FALSE    Cleveland          Barons        Cleveland Barons      Cleveland        19761977            1976-77       19771978           1977-78                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/CLE_19761977-19771978_light.svg
#> 50:           21      47               AFM       FALSE      Atlanta          Flames          Atlanta Flames        Atlanta        19721973            1972-73       19791980           1979-80                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/AFM_19721973-19791980_light.svg
#> 51:           23      35               CLR       FALSE     Colorado         Rockies        Colorado Rockies       Colorado        19761977            1976-77       19811982           1981-82                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/CLR_19761977-19811982_light.svg
#> 52:           15      31               MNS       FALSE    Minnesota     North Stars   Minnesota North Stars      Minnesota        19671968            1967-68       19921993           1992-93                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/MNS_19911992-19921993_light.svg
#> 53:           27      32               QUE       FALSE       Quebec       Nordiques        Quebec Nordiques         Quebec        19791980            1979-80       19941995           1994-95                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/QUE_19791980-19941995_light.svg
#> 54:           28      33               WIN       FALSE     Winnipeg     Jets (1979)    Winnipeg Jets (1979)       Winnipeg        19791980            1979-80       19951996           1995-96                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/WIN_19901991-19951996_light.svg
#> 55:           26      34               HFD       FALSE     Hartford         Whalers        Hartford Whalers       Hartford        19791980            1979-80       19961997           1996-97                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/HFD_19921993-19961997_light.svg
#> 56:           35      11               ATL       FALSE      Atlanta       Thrashers       Atlanta Thrashers        Atlanta        19992000            1999-00       20102011           2010-11                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/ATL_19992000-20102011_light.svg
#> 57:           28      27               PHX       FALSE      Phoenix         Coyotes         Phoenix Coyotes        Phoenix        19961997            1996-97       20132014           2013-14                   NA                           <NA>                   <NA>                 NA                         <NA>                 <NA>                     <NA>                 <NA>                       <NA>              <NA>                <NA>               NA                NA https://assets.nhle.com/logos/nhl/svg/PHX_20032004-20132014_light.svg
#>     franchise_id team_id team_abbreviation team_active   team_place       team_name           team_fullname team_shortname season_first_id season_first_years season_last_id season_last_years conference_active_id conference_active_abbreviation conference_active_name division_active_id division_active_abbreviation division_active_name        venue_active_name venue_active_country venue_active_stateprovince venue_active_city     venue_active_tz venue_active_lat venue_active_long                                                         logo_last_url