Get a tidy dataset of NHL players game logs

The function tidy_players_gamelogs() is meant to be a user-friendly way of getting the NHL game logs of selected players.

tidy_players_gamelogs(
  players_id,
  seasons_id = NULL,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = NULL
)

Arguments

players_id
Integer vector indicating the NHL ID of players for whom the game logs will be returned.

seasons_id
(optional) Character vector indicating the season for which the game logs will be returned. The required format for each character is 'xxxxyyyy'. Default to NULL which will return the full career game logs for each player.

regular
(optional) Logical indicating if the regular season schedule should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs schedule should be returned. Default to TRUE.

tz
(optional) Character specifying the timezone that should be used for datetime. Default to the user system timezone.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Get the NHL career game logs of Vincent Lecavalier
head(tidy_players_gamelogs(8467329L))

#>          player_name season_years season_type       game_datetime
#> 1 Vincent Lecavalier      1998-99     regular 1998-10-09 19:30:00
#> 2 Vincent Lecavalier      1998-99     regular 1998-10-10 19:00:00
#> 3 Vincent Lecavalier      1998-99     regular 1998-10-14 19:00:00
#> 4 Vincent Lecavalier      1998-99     regular 1998-10-16 19:00:00
#> 5 Vincent Lecavalier      1998-99     regular 1998-10-18 19:00:00
#> 6 Vincent Lecavalier      1998-99     regular 1998-10-21 19:00:00
#>                   venue_name team_status team_abbreviation team_score
#> 1 National Car Rental Center        away               TBL          1
#> 2        Greensboro Coliseum        away               TBL          4
#> 3                 Ice Palace        home               TBL          0
#> 4                 Ice Palace        home               TBL          2
#> 5                 Ice Palace        home               TBL          1
#> 6                 Ice Palace        home               TBL          5
#>   opponent_score opponent_abbreviation game_status game_nbot game_shootout
#> 1              4                   FLA       final         0         FALSE
#> 2              4                   CAR       final         1         FALSE
#> 3              2                   NYI       final         0         FALSE
#> 4              5                   PHI       final         0         FALSE
#> 5              4                   WSH       final         0         FALSE
#> 6              0                   PIT       final         0         FALSE
#>   skater_games skater_goals skater_assists skater_points skater_plusminus
#> 1            1            0              0             0               -2
#> 2            1            0              0             0                0
#> 3            1            0              0             0               -1
#> 4            1            0              0             0               -1
#> 5            1            0              0             0                0
#> 6            1            0              1             1                0
#>   skater_pim skater_toi skater_shifts skater_gwg skater_otg skater_shots
#> 1          0  15.383333            18          0          0            2
#> 2          0  14.783333            18          0          0            2
#> 3          0  16.583333            19          0          0            1
#> 4          0  13.450000            15          0          0            3
#> 5          0   8.466667             7          0          0            0
#> 6          0  10.150000            11          0          0            1
#>   skater_blocked skater_hits skater_ev_goals skater_ev_assists skater_ev_points
#> 1              0           0               0                 0                0
#> 2              0           1               0                 0                0
#> 3              1           2               0                 0                0
#> 4              0           2               0                 0                0
#> 5              0           0               0                 0                0
#> 6              0           0               0                 0                0
#>   skater_ev_toi skater_pp_goals skater_pp_assists skater_pp_points
#> 1     10.783333               0                 0                0
#> 2     12.850000               0                 0                0
#> 3     14.033333               0                 0                0
#> 4      9.483333               0                 0                0
#> 5      6.533333               0                 0                0
#> 6      9.233333               0                 1                1
#>   skater_pp_toi skater_pk_goals skater_pk_assists skater_pk_points
#> 1     2.6000000               0                 0                0
#> 2     1.7500000               0                 0                0
#> 3     2.0000000               0                 0                0
#> 4     3.6333333               0                 0                0
#> 5     1.8666667               0                 0                0
#> 6     0.7666667               0                 0                0
#>   skater_pk_toi goalie_games goalie_started goalie_wins goalie_losses
#> 1    2.00000000           NA             NA          NA            NA
#> 2    0.18333333           NA             NA          NA            NA
#> 3    0.55000000           NA             NA          NA            NA
#> 4    0.33333333           NA             NA          NA            NA
#> 5    0.06666667           NA             NA          NA            NA
#> 6    0.15000000           NA             NA          NA            NA
#>   goalie_ties goalie_ot goalie_shutouts goalie_shotagainst goalie_goalagainst
#> 1          NA        NA              NA                 NA                 NA
#> 2          NA        NA              NA                 NA                 NA
#> 3          NA        NA              NA                 NA                 NA
#> 4          NA        NA              NA                 NA                 NA
#> 5          NA        NA              NA                 NA                 NA
#> 6          NA        NA              NA                 NA                 NA
#>   goalie_savepct goalie_gaa goalie_toi goalie_ev_shotagainst
#> 1             NA         NA         NA                    NA
#> 2             NA         NA         NA                    NA
#> 3             NA         NA         NA                    NA
#> 4             NA         NA         NA                    NA
#> 5             NA         NA         NA                    NA
#> 6             NA         NA         NA                    NA
#>   goalie_ev_goalagainst goalie_ev_savepct goalie_pp_shotagainst
#> 1                    NA                NA                    NA
#> 2                    NA                NA                    NA
#> 3                    NA                NA                    NA
#> 4                    NA                NA                    NA
#> 5                    NA                NA                    NA
#> 6                    NA                NA                    NA
#>   goalie_pp_goalagainst goalie_pp_savepct goalie_pk_shotagainst
#> 1                    NA                NA                    NA
#> 2                    NA                NA                    NA
#> 3                    NA                NA                    NA
#> 4                    NA                NA                    NA
#> 5                    NA                NA                    NA
#> 6                    NA                NA                    NA
#>   goalie_pk_goalagainst goalie_pk_savepct
#> 1                    NA                NA
#> 2                    NA                NA
#> 3                    NA                NA
#> 4                    NA                NA
#> 5                    NA                NA
#> 6                    NA                NA

# Get the NHL 2010-11 playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
tidy_players_gamelogs(
  players_id = c(8467329L, 8471679L),
  seasons_id = "20102011",
  regular    = FALSE,
  keep_id    = TRUE
)

#>    player_id        player_name season_id season_years season_type    game_id
#> 1    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030141
#> 2    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030142
#> 3    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030143
#> 4    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030144
#> 5    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030145
#> 6    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030146
#> 7    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030147
#> 8    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030211
#> 9    8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030212
#> 10   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030213
#> 11   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030214
#> 12   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030311
#> 13   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030312
#> 14   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030313
#> 15   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030314
#> 16   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030315
#> 17   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030316
#> 18   8467329 Vincent Lecavalier  20102011      2010-11    playoffs 2010030317
#> 19   8471679        Carey Price  20102011      2010-11    playoffs 2010030131
#> 20   8471679        Carey Price  20102011      2010-11    playoffs 2010030132
#> 21   8471679        Carey Price  20102011      2010-11    playoffs 2010030133
#> 22   8471679        Carey Price  20102011      2010-11    playoffs 2010030134
#> 23   8471679        Carey Price  20102011      2010-11    playoffs 2010030135
#> 24   8471679        Carey Price  20102011      2010-11    playoffs 2010030136
#> 25   8471679        Carey Price  20102011      2010-11    playoffs 2010030137
#>          game_datetime           venue_name team_status team_id
#> 1  2011-04-13 19:00:00 CONSOL Energy Center        away      14
#> 2  2011-04-15 19:00:00 CONSOL Energy Center        away      14
#> 3  2011-04-18 19:30:00 St. Pete Times Forum        home      14
#> 4  2011-04-20 19:00:00 St. Pete Times Forum        home      14
#> 5  2011-04-23 12:00:00 CONSOL Energy Center        away      14
#> 6  2011-04-25 19:00:00 St. Pete Times Forum        home      14
#> 7  2011-04-27 20:00:00 CONSOL Energy Center        away      14
#> 8  2011-04-29 19:00:00       Verizon Center        away      14
#> 9  2011-05-01 19:00:00       Verizon Center        away      14
#> 10 2011-05-03 18:30:00 St. Pete Times Forum        home      14
#> 11 2011-05-04 19:00:00 St. Pete Times Forum        home      14
#> 12 2011-05-14 20:00:00            TD Garden        away      14
#> 13 2011-05-17 20:00:00            TD Garden        away      14
#> 14 2011-05-19 20:00:00 St. Pete Times Forum        home      14
#> 15 2011-05-21 13:30:00 St. Pete Times Forum        home      14
#> 16 2011-05-23 20:00:00            TD Garden        away      14
#> 17 2011-05-25 20:00:00 St. Pete Times Forum        home      14
#> 18 2011-05-27 20:00:00            TD Garden        away      14
#> 19 2011-04-14 19:00:00            TD Garden        away       8
#> 20 2011-04-16 19:00:00            TD Garden        away       8
#> 21 2011-04-18 19:30:00          Centre Bell        home       8
#> 22 2011-04-21 19:00:00          Centre Bell        home       8
#> 23 2011-04-23 19:00:00            TD Garden        away       8
#> 24 2011-04-26 19:00:00          Centre Bell        home       8
#> 25 2011-04-27 19:00:00            TD Garden        away       8
#>    team_abbreviation team_score opponent_score opponent_abbreviation
#> 1                TBL          0              3                   PIT
#> 2                TBL          5              1                   PIT
#> 3                TBL          2              3                   PIT
#> 4                TBL          2              3                   PIT
#> 5                TBL          8              2                   PIT
#> 6                TBL          4              2                   PIT
#> 7                TBL          1              0                   PIT
#> 8                TBL          4              2                   WSH
#> 9                TBL          3              2                   WSH
#> 10               TBL          4              3                   WSH
#> 11               TBL          5              3                   WSH
#> 12               TBL          5              2                   BOS
#> 13               TBL          5              6                   BOS
#> 14               TBL          0              2                   BOS
#> 15               TBL          5              3                   BOS
#> 16               TBL          1              3                   BOS
#> 17               TBL          5              4                   BOS
#> 18               TBL          0              1                   BOS
#> 19               MTL          2              0                   BOS
#> 20               MTL          3              1                   BOS
#> 21               MTL          2              4                   BOS
#> 22               MTL          4              5                   BOS
#> 23               MTL          1              2                   BOS
#> 24               MTL          2              1                   BOS
#> 25               MTL          3              4                   BOS
#>    opponent_id game_status game_nbot game_shootout skater_games skater_goals
#> 1            5       final         0         FALSE            1            0
#> 2            5       final         0         FALSE            1            1
#> 3            5       final         0         FALSE            1            0
#> 4            5       final         2         FALSE            1            0
#> 5            5       final         0         FALSE            1            1
#> 6            5       final         0         FALSE            1            0
#> 7            5       final         0         FALSE            1            0
#> 8           15       final         0         FALSE            1            0
#> 9           15       final         1         FALSE            1            2
#> 10          15       final         0         FALSE            1            1
#> 11          15       final         0         FALSE            1            0
#> 12           6       final         0         FALSE            1            0
#> 13           6       final         0         FALSE            1            1
#> 14           6       final         0         FALSE            1            0
#> 15           6       final         0         FALSE            1            0
#> 16           6       final         0         FALSE            1            0
#> 17           6       final         0         FALSE            1            0
#> 18           6       final         0         FALSE            1            0
#> 19           6       final         0         FALSE           NA           NA
#> 20           6       final         0         FALSE           NA           NA
#> 21           6       final         0         FALSE           NA           NA
#> 22           6       final         1         FALSE           NA           NA
#> 23           6       final         2         FALSE           NA           NA
#> 24           6       final         0         FALSE           NA           NA
#> 25           6       final         1         FALSE           NA           NA
#>    skater_assists skater_points skater_plusminus skater_pim skater_toi
#> 1               0             0               -2          0   16.83333
#> 2               0             1                0          0   18.40000
#> 3               1             1                0          2   23.65000
#> 4               1             1                0          2   26.46667
#> 5               1             2                2          4   17.46667
#> 6               1             1                1          0   18.18333
#> 7               0             0                0          0   17.71667
#> 8               2             2                0          0   21.73333
#> 9               0             2                2          0   19.68333
#> 10              0             1                1          0   19.50000
#> 11              1             1                0          0   20.75000
#> 12              0             0                3          2   18.15000
#> 13              3             4                1          2   20.10000
#> 14              0             0               -1          0   23.15000
#> 15              1             1                0          0   18.80000
#> 16              0             0               -1          2   19.50000
#> 17              2             2                0          2   17.65000
#> 18              0             0                0          0   19.76667
#> 19             NA            NA               NA         NA         NA
#> 20             NA            NA               NA         NA         NA
#> 21             NA            NA               NA         NA         NA
#> 22             NA            NA               NA         NA         NA
#> 23             NA            NA               NA         NA         NA
#> 24             NA            NA               NA         NA         NA
#> 25             NA            NA               NA         NA         NA
#>    skater_shifts skater_gwg skater_otg skater_shots skater_blocked skater_hits
#> 1             22          0          0            4              1           4
#> 2             24          1          0            1              0           0
#> 3             26          0          0            5              0           0
#> 4             34          0          0            4              0           3
#> 5             25          1          0            2              1           0
#> 6             25          0          0            1              1           1
#> 7             22          0          0            1              1           2
#> 8             25          0          0            2              1           2
#> 9             24          1          1            3              0           0
#> 10            25          0          0            4              0           1
#> 11            26          0          0            3              0           3
#> 12            23          0          0            5              0           0
#> 13            26          0          0            5              1           0
#> 14            23          0          0            5              0           1
#> 15            24          0          0            5              0           1
#> 16            25          0          0            4              0           0
#> 17            20          0          0            0              0           1
#> 18            25          0          0            2              1           1
#> 19            NA         NA         NA           NA             NA          NA
#> 20            NA         NA         NA           NA             NA          NA
#> 21            NA         NA         NA           NA             NA          NA
#> 22            NA         NA         NA           NA             NA          NA
#> 23            NA         NA         NA           NA             NA          NA
#> 24            NA         NA         NA           NA             NA          NA
#> 25            NA         NA         NA           NA             NA          NA
#>    skater_ev_goals skater_ev_assists skater_ev_points skater_ev_toi
#> 1                0                 0                0      14.90000
#> 2                0                 0                0      14.05000
#> 3                0                 0                0      18.76667
#> 4                0                 1                1      20.43333
#> 5                1                 1                2      13.06667
#> 6                0                 1                1      13.53333
#> 7                0                 0                0      14.78333
#> 8                0                 1                1      16.83333
#> 9                1                 0                1      12.80000
#> 10               1                 0                1      15.03333
#> 11               0                 0                0      13.88333
#> 12               0                 0                0      14.08333
#> 13               0                 3                3      16.01667
#> 14               0                 0                0      18.78333
#> 15               0                 1                1      16.13333
#> 16               0                 0                0      14.95000
#> 17               0                 1                1      13.31667
#> 18               0                 0                0      19.76667
#> 19              NA                NA               NA            NA
#> 20              NA                NA               NA            NA
#> 21              NA                NA               NA            NA
#> 22              NA                NA               NA            NA
#> 23              NA                NA               NA            NA
#> 24              NA                NA               NA            NA
#> 25              NA                NA               NA            NA
#>    skater_pp_goals skater_pp_assists skater_pp_points skater_pp_toi
#> 1                0                 0                0      1.283333
#> 2                1                 0                1      3.800000
#> 3                0                 1                1      4.883333
#> 4                0                 0                0      5.916667
#> 5                0                 0                0      4.400000
#> 6                0                 0                0      4.650000
#> 7                0                 0                0      2.466667
#> 8                0                 1                1      4.316667
#> 9                1                 0                1      6.566667
#> 10               0                 0                0      4.333333
#> 11               0                 1                1      6.866667
#> 12               0                 0                0      4.066667
#> 13               1                 0                1      3.783333
#> 14               0                 0                0      4.083333
#> 15               0                 0                0      2.666667
#> 16               0                 0                0      4.166667
#> 17               0                 1                1      4.100000
#> 18               0                 0                0      0.000000
#> 19              NA                NA               NA            NA
#> 20              NA                NA               NA            NA
#> 21              NA                NA               NA            NA
#> 22              NA                NA               NA            NA
#> 23              NA                NA               NA            NA
#> 24              NA                NA               NA            NA
#> 25              NA                NA               NA            NA
#>    skater_pk_goals skater_pk_assists skater_pk_points skater_pk_toi
#> 1                0                 0                0     0.6500000
#> 2                0                 0                0     0.5500000
#> 3                0                 0                0     0.0000000
#> 4                0                 0                0     0.1166667
#> 5                0                 0                0     0.0000000
#> 6                0                 0                0     0.0000000
#> 7                0                 0                0     0.4666667
#> 8                0                 0                0     0.5833333
#> 9                0                 0                0     0.3166667
#> 10               0                 0                0     0.1333333
#> 11               0                 0                0     0.0000000
#> 12               0                 0                0     0.0000000
#> 13               0                 0                0     0.3000000
#> 14               0                 0                0     0.2833333
#> 15               0                 0                0     0.0000000
#> 16               0                 0                0     0.3833333
#> 17               0                 0                0     0.2333333
#> 18               0                 0                0     0.0000000
#> 19              NA                NA               NA            NA
#> 20              NA                NA               NA            NA
#> 21              NA                NA               NA            NA
#> 22              NA                NA               NA            NA
#> 23              NA                NA               NA            NA
#> 24              NA                NA               NA            NA
#> 25              NA                NA               NA            NA
#>    goalie_games goalie_started goalie_wins goalie_losses goalie_ties goalie_ot
#> 1            NA             NA          NA            NA          NA        NA
#> 2            NA             NA          NA            NA          NA        NA
#> 3            NA             NA          NA            NA          NA        NA
#> 4            NA             NA          NA            NA          NA        NA
#> 5            NA             NA          NA            NA          NA        NA
#> 6            NA             NA          NA            NA          NA        NA
#> 7            NA             NA          NA            NA          NA        NA
#> 8            NA             NA          NA            NA          NA        NA
#> 9            NA             NA          NA            NA          NA        NA
#> 10           NA             NA          NA            NA          NA        NA
#> 11           NA             NA          NA            NA          NA        NA
#> 12           NA             NA          NA            NA          NA        NA
#> 13           NA             NA          NA            NA          NA        NA
#> 14           NA             NA          NA            NA          NA        NA
#> 15           NA             NA          NA            NA          NA        NA
#> 16           NA             NA          NA            NA          NA        NA
#> 17           NA             NA          NA            NA          NA        NA
#> 18           NA             NA          NA            NA          NA        NA
#> 19            1              1           1             0           0         0
#> 20            1              1           1             0           0         0
#> 21            1              1           0             1           0         0
#> 22            1              1           0             1           0         0
#> 23            1              1           0             1           0         0
#> 24            1              1           1             0           0         0
#> 25            1              1           0             1           0         0
#>    goalie_shutouts goalie_shotagainst goalie_goalagainst goalie_savepct
#> 1               NA                 NA                 NA             NA
#> 2               NA                 NA                 NA             NA
#> 3               NA                 NA                 NA             NA
#> 4               NA                 NA                 NA             NA
#> 5               NA                 NA                 NA             NA
#> 6               NA                 NA                 NA             NA
#> 7               NA                 NA                 NA             NA
#> 8               NA                 NA                 NA             NA
#> 9               NA                 NA                 NA             NA
#> 10              NA                 NA                 NA             NA
#> 11              NA                 NA                 NA             NA
#> 12              NA                 NA                 NA             NA
#> 13              NA                 NA                 NA             NA
#> 14              NA                 NA                 NA             NA
#> 15              NA                 NA                 NA             NA
#> 16              NA                 NA                 NA             NA
#> 17              NA                 NA                 NA             NA
#> 18              NA                 NA                 NA             NA
#> 19               1                 31                  0      1.0000000
#> 20               0                 35                  1      0.9714286
#> 21               0                 24                  3      0.8750000
#> 22               0                 35                  5      0.8571429
#> 23               0                 51                  2      0.9607843
#> 24               0                 32                  1      0.9687500
#> 25               0                 34                  4      0.8823529
#>    goalie_gaa goalie_toi goalie_ev_shotagainst goalie_ev_goalagainst
#> 1          NA         NA                    NA                    NA
#> 2          NA         NA                    NA                    NA
#> 3          NA         NA                    NA                    NA
#> 4          NA         NA                    NA                    NA
#> 5          NA         NA                    NA                    NA
#> 6          NA         NA                    NA                    NA
#> 7          NA         NA                    NA                    NA
#> 8          NA         NA                    NA                    NA
#> 9          NA         NA                    NA                    NA
#> 10         NA         NA                    NA                    NA
#> 11         NA         NA                    NA                    NA
#> 12         NA         NA                    NA                    NA
#> 13         NA         NA                    NA                    NA
#> 14         NA         NA                    NA                    NA
#> 15         NA         NA                    NA                    NA
#> 16         NA         NA                    NA                    NA
#> 17         NA         NA                    NA                    NA
#> 18         NA         NA                    NA                    NA
#> 19   0.000000   60.00000                    26                     0
#> 20   1.000278   59.98333                    29                     1
#> 21   3.033708   59.33333                    20                     3
#> 22   4.840011   61.98333                    35                     5
#> 23   1.352113   88.75000                    46                     2
#> 24   1.000000   60.00000                    28                     1
#> 25   3.667855   65.43333                    32                     4
#>    goalie_ev_savepct goalie_pp_shotagainst goalie_pp_goalagainst
#> 1                 NA                    NA                    NA
#> 2                 NA                    NA                    NA
#> 3                 NA                    NA                    NA
#> 4                 NA                    NA                    NA
#> 5                 NA                    NA                    NA
#> 6                 NA                    NA                    NA
#> 7                 NA                    NA                    NA
#> 8                 NA                    NA                    NA
#> 9                 NA                    NA                    NA
#> 10                NA                    NA                    NA
#> 11                NA                    NA                    NA
#> 12                NA                    NA                    NA
#> 13                NA                    NA                    NA
#> 14                NA                    NA                    NA
#> 15                NA                    NA                    NA
#> 16                NA                    NA                    NA
#> 17                NA                    NA                    NA
#> 18                NA                    NA                    NA
#> 19         1.0000000                     5                     0
#> 20         0.9655172                     5                     0
#> 21         0.8500000                     3                     0
#> 22         0.8571429                     0                     0
#> 23         0.9565217                     4                     0
#> 24         0.9642857                     2                     0
#> 25         0.8750000                     2                     0
#>    goalie_pp_savepct goalie_pk_shotagainst goalie_pk_goalagainst
#> 1                 NA                    NA                    NA
#> 2                 NA                    NA                    NA
#> 3                 NA                    NA                    NA
#> 4                 NA                    NA                    NA
#> 5                 NA                    NA                    NA
#> 6                 NA                    NA                    NA
#> 7                 NA                    NA                    NA
#> 8                 NA                    NA                    NA
#> 9                 NA                    NA                    NA
#> 10                NA                    NA                    NA
#> 11                NA                    NA                    NA
#> 12                NA                    NA                    NA
#> 13                NA                    NA                    NA
#> 14                NA                    NA                    NA
#> 15                NA                    NA                    NA
#> 16                NA                    NA                    NA
#> 17                NA                    NA                    NA
#> 18                NA                    NA                    NA
#> 19                 1                     0                     0
#> 20                 1                     1                     0
#> 21                 1                     1                     0
#> 22               NaN                     0                     0
#> 23                 1                     1                     0
#> 24                 1                     2                     0
#> 25                 1                     0                     0
#>    goalie_pk_savepct
#> 1                 NA
#> 2                 NA
#> 3                 NA
#> 4                 NA
#> 5                 NA
#> 6                 NA
#> 7                 NA
#> 8                 NA
#> 9                 NA
#> 10                NA
#> 11                NA
#> 12                NA
#> 13                NA
#> 14                NA
#> 15                NA
#> 16                NA
#> 17                NA
#> 18                NA
#> 19               NaN
#> 20                 1
#> 21                 1
#> 22               NaN
#> 23                 1
#> 24                 1
#> 25               NaN