Get a tidy dataset of NHL players metadata

The function tidy_players_meta() is meant to be a user-friendly way of getting metadata about the selected players.

tidy_players_meta(players_id, keep_id = FALSE, return_datatable = NULL)

Arguments

players_id
Integer vector indicating the NHL ID of players for whom the game logs will be returned.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Load the data.table package to easily manipulate the data
library(data.table)

# Get metadata about Vincent Lecavalier
metadata_lecavalier <- tidy_players_meta(8467329L)

# Print the column names
colnames(metadata_lecavalier)

#>  [1] "player_name"                "player_birth_date"         
#>  [3] "player_birth_city"          "player_birth_stateprovince"
#>  [5] "player_birth_country"       "player_nationality"        
#>  [7] "player_inches"              "player_pounds"             
#>  [9] "player_hand"                "player_position"           
#> [11] "player_position_type"      

# Print an excerpt of the data
metadata_lecavalier[, .(player_name, player_birth_date, player_nationality, player_hand)]

#>           player_name player_birth_date player_nationality player_hand
#> 1: Vincent Lecavalier        1980-04-21                CAN           L

# Get metadata about both Vincent Lecavalier and Carey Price, keeping the IDs
metadata_lecavalier_price <- tidy_players_meta(
  players_id = c(8467329L, 8471679L),
  keep_id    = TRUE
)

# Print the column names
colnames(metadata_lecavalier_price)

#>  [1] "player_id"                  "player_name"               
#>  [3] "player_birth_date"          "player_birth_city"         
#>  [5] "player_birth_stateprovince" "player_birth_country"      
#>  [7] "player_nationality"         "player_inches"             
#>  [9] "player_pounds"              "player_hand"               
#> [11] "player_position"            "player_position_type"      

# Print an excerpt of the data
metadata_lecavalier_price[, .(player_id, player_name, player_inches, player_pounds, player_position)]

#>    player_id        player_name player_inches player_pounds player_position
#> 1:   8467329 Vincent Lecavalier            76           215               C
#> 2:   8471679        Carey Price            75           217               G