Get a tidy dataset of NHL players statistics

The function tidy_players_stats() is meant to be a user-friendly way of getting the NHL year-by-year statistics of selected players.

tidy_players_stats(
  players_id,
  regular = TRUE,
  playoffs = TRUE,
  keep_id = FALSE,
  return_datatable = NULL
)

Arguments

players_id
Integer vector indicating the NHL ID of players for whom the statistics will be returned.

regular
(optional) Logical indicating if the regular season statistics should be returned. Default to TRUE.

playoffs
(optional) Logical indicating if the playoffs statistics should be returned. Default to TRUE.

keep_id
(optional) Logical indicating if the IDs of different dimensions should be returned. Default to FALSE.

return_datatable
(optional) Logical indicating whether or not a data.table should be returned. Default to TRUE if the data.table package is attached in the active session.

Examples

# Load the data.table package to easily manipulate the data
library(data.table)

# Get the NHL career statistics of Vincent Lecavalier
stats_lecavalier <- tidy_players_stats(8467329L)

# Print the column names
colnames(stats_lecavalier)

#>  [1] "player_name"           "season_years"          "season_type"          
#>  [4] "team_abbreviation"     "skater_games"          "skater_goals"         
#>  [7] "skater_assists"        "skater_points"         "skater_plusminus"     
#> [10] "skater_pim"            "skater_toi"            "skater_shifts"        
#> [13] "skater_gwg"            "skater_otg"            "skater_shots"         
#> [16] "skater_blocked"        "skater_hits"           "skater_ev_goals"      
#> [19] "skater_ev_assists"     "skater_ev_points"      "skater_ev_toi"        
#> [22] "skater_pp_goals"       "skater_pp_assists"     "skater_pp_points"     
#> [25] "skater_pp_toi"         "skater_pk_goals"       "skater_pk_assists"    
#> [28] "skater_pk_points"      "skater_pk_toi"         "goalie_games"         
#> [31] "goalie_started"        "goalie_wins"           "goalie_losses"        
#> [34] "goalie_ties"           "goalie_ot"             "goalie_shutouts"      
#> [37] "goalie_shotagainst"    "goalie_goalagainst"    "goalie_savepct"       
#> [40] "goalie_gaa"            "goalie_toi"            "goalie_ev_shotagainst"
#> [43] "goalie_ev_goalagainst" "goalie_ev_savepct"     "goalie_pp_shotagainst"
#> [46] "goalie_pp_goalagainst" "goalie_pp_savepct"     "goalie_pk_shotagainst"
#> [49] "goalie_pk_goalagainst" "goalie_pk_savepct"    

# Print an excerpt of the data
stats_lecavalier[, .(season_years, season_type, team_abbreviation, skater_games, skater_points)]

#>     season_years season_type team_abbreviation skater_games skater_points
#>  1:      1998-99     regular               TBL           82            28
#>  2:      1999-00     regular               TBL           80            67
#>  3:      2000-01     regular               TBL           68            51
#>  4:      2001-02     regular               TBL           76            37
#>  5:      2002-03     regular               TBL           80            78
#>  6:      2002-03    playoffs               TBL           11             6
#>  7:      2003-04     regular               TBL           81            66
#>  8:      2003-04    playoffs               TBL           23            16
#>  9:      2005-06     regular               TBL           80            75
#> 10:      2005-06    playoffs               TBL            5             4
#> 11:      2006-07     regular               TBL           82           108
#> 12:      2006-07    playoffs               TBL            6             7
#> 13:      2007-08     regular               TBL           81            92
#> 14:      2008-09     regular               TBL           77            67
#> 15:      2009-10     regular               TBL           82            70
#> 16:      2010-11     regular               TBL           65            54
#> 17:      2010-11    playoffs               TBL           18            19
#> 18:      2011-12     regular               TBL           64            49
#> 19:      2012-13     regular               TBL           39            32
#> 20:      2013-14     regular               PHI           69            37
#> 21:      2013-14    playoffs               PHI            7             2
#> 22:      2014-15     regular               PHI           57            20
#> 23:      2015-16     regular               PHI            7             1
#> 24:      2015-16     regular               LAK           42            17
#> 25:      2015-16    playoffs               LAK            5             2
#>     season_years season_type team_abbreviation skater_games skater_points

# Get the NHL career playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
stats_playoffs_lecavalier_price <- tidy_players_stats(
  players_id = c(8467329L, 8471679L),
  regular    = FALSE,
  keep_id    = TRUE
)

# Print the column names
colnames(stats_playoffs_lecavalier_price)

#>  [1] "player_id"             "player_name"           "season_id"            
#>  [4] "season_years"          "season_type"           "team_id"              
#>  [7] "team_abbreviation"     "skater_games"          "skater_goals"         
#> [10] "skater_assists"        "skater_points"         "skater_plusminus"     
#> [13] "skater_pim"            "skater_toi"            "skater_shifts"        
#> [16] "skater_gwg"            "skater_otg"            "skater_shots"         
#> [19] "skater_blocked"        "skater_hits"           "skater_ev_goals"      
#> [22] "skater_ev_assists"     "skater_ev_points"      "skater_ev_toi"        
#> [25] "skater_pp_goals"       "skater_pp_assists"     "skater_pp_points"     
#> [28] "skater_pp_toi"         "skater_pk_goals"       "skater_pk_assists"    
#> [31] "skater_pk_points"      "skater_pk_toi"         "goalie_games"         
#> [34] "goalie_started"        "goalie_wins"           "goalie_losses"        
#> [37] "goalie_ties"           "goalie_ot"             "goalie_shutouts"      
#> [40] "goalie_shotagainst"    "goalie_goalagainst"    "goalie_savepct"       
#> [43] "goalie_gaa"            "goalie_toi"            "goalie_ev_shotagainst"
#> [46] "goalie_ev_goalagainst" "goalie_ev_savepct"     "goalie_pp_shotagainst"
#> [49] "goalie_pp_goalagainst" "goalie_pp_savepct"     "goalie_pk_shotagainst"
#> [52] "goalie_pk_goalagainst" "goalie_pk_savepct"    

# Print an excerpt of the data
stats_playoffs_lecavalier_price[, .(player_id, player_name, season_years, skater_points, goalie_savepct)]

#>     player_id        player_name season_years skater_points goalie_savepct
#>  1:   8467329 Vincent Lecavalier      2002-03             6             NA
#>  2:   8467329 Vincent Lecavalier      2003-04            16             NA
#>  3:   8467329 Vincent Lecavalier      2005-06             4             NA
#>  4:   8467329 Vincent Lecavalier      2006-07             7             NA
#>  5:   8467329 Vincent Lecavalier      2010-11            19             NA
#>  6:   8467329 Vincent Lecavalier      2013-14             2             NA
#>  7:   8467329 Vincent Lecavalier      2015-16             2             NA
#>  8:   8471679        Carey Price      2007-08            NA      0.9013158
#>  9:   8471679        Carey Price      2008-09            NA      0.8780488
#> 10:   8471679        Carey Price      2009-10            NA      0.8904110
#> 11:   8471679        Carey Price      2010-11            NA      0.9338843
#> 12:   8471679        Carey Price      2012-13            NA      0.8943089
#> 13:   8471679        Carey Price      2013-14            NA      0.9189944
#> 14:   8471679        Carey Price      2014-15            NA      0.9204545
#> 15:   8471679        Carey Price      2016-17            NA      0.9329609
#> 16:   8471679        Carey Price      2019-20            NA      0.9361702